% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rimplementation.R
\name{aboveground}
\alias{aboveground}
\title{Estimate temperature and humidity at specified height above ground}
\usage{
aboveground(micro, reqhgt = 0.05, pai_a = NA, tfact = 1.5)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{reqhgt}{height above ground for which wind speeds are wanted. If negative (below ground) wind friction velocity only is returned}

\item{pai_a}{plant area index above `reqhgt`. Determined from total `pai` if not supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}
}
\value{
a list of the following:
\describe{
  \item{Tz}{Air temperatures at height `reqhgt` (deg C)}
  \item{tleaf}{Leaf temperatures at height `reqhgt` (deg C)}
  \item{T0}{Ground surface temperatures (deg C)}
  \item{relhum}{Relative humidities at height `reqhgt` (percentage)}
  \item{windspeed}{Wind speeds at height `reqhgt` (m/s)}
  \item{Rdirdown}{Flux density of downward direct radiation at `reqhgt` (W/m^2 - on the horizontal}
  \item{Rdirdown}{Flux density of downward diffuse radiation at `reqhgt` (W/m^2)}
  \item{Rlwdown}{Flux density of downward longwave radiation at `reqhgt` (W/m^2)}
  \item{Rswup}{Flux density of upward shorwtave radiation at `reqhgt` (W/m^2), assumed diffuse}
  \item{Rlwup}{Flux density of downward longwave radiation at `reqhgt` (W/m^2)}
}
}
\description{
The function `aboveground` runs the above ground component of
the microclimate model.
}
\details{
`pai_a` is used to calculate the radiation intercepted by leaves at `reqhgt` if
below canopy. If not supplied it is calculated from total plant area index by
assuming leaf density within the canopy is uniformly vertically distributed. If supplied
it must have the same dimensions as micro$pai. I.e. with the same x and y dims as the
the supplied dtm and values for each hour as the z dimension.
}
\seealso{
[belowground()]
}
