% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albedo}
\alias{albedo}
\title{White-sky albedo at Caerthillian Cove}
\format{
A 50 x 50 cell \code{PackedSpatRaster}.
}
\source{
Derived from aerial imagery available through \url{https://digimap.edina.ac.uk/}.
}
\usage{
albedo
}
\description{
One-metre-resolution white-sky albedo for a 50 x 50 m area of Caerthillian
Cove, Cornwall, on the British National Grid (EPSG:27700; extent 169475--169525
E, 12475--12525 N).
}
\keyword{datasets}
