% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rimplementation.R
\name{belowground}
\alias{belowground}
\title{Estimate temperature at specified height below ground}
\usage{
belowground(micro, reqhgt = -0.05, pai_a = NA, tfact = 1.5)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{reqhgt}{height above ground for which wind speeds are wanted. If negative (below ground) wind friction velocity only is returned}

\item{pai_a}{plant area index above `reqhgt`. Determined from total `pai` if not supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}
}
\value{
a list of the following:
\describe{
  \item{Tz}{Soil temperatures at height `reqhgt` below ground (deg C)}
  \item{T0}{Ground surface temperatures (deg C)}
  \item{soilm}{Soil moisture in top 10 cm of soil (volumetric water fraction)}
}
}
\description{
The function `belowground` runs the below ground component of
the microclimate model.
}
\details{
`pai_a` is used to calculate the radiation intercepted by leaves at `reqhgt` if
below canopy. If not supplied it is calculated from total plant area index by
assuming leaf density within the canopy is uniformly vertically distributed. If supplied
it must have the same dimensions as micro$pai. I.e. with the same x and y dims as the
the supplied dtm and values for each hour as the z dimension.
}
\seealso{
[aboveground()]
}
