% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Cafromyear}
\alias{Cafromyear}
\title{Estimate atmospheric CO2 concentration from year}
\usage{
Cafromyear(year)
}
\arguments{
\item{year}{Calendar year (1750-2100).}
}
\value{
Estimated atmospheric CO2 concentration (ppm).
}
\description{
Estimates atmospheric CO2 concentration from calendar year using piecewise
polynomial fits spanning the historical and projection periods used by the
model.
}
