% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climdata}
\alias{climdata}
\title{Example hourly weather from Caerthillian Cove}
\format{
A data frame with columns:
\describe{
  \item{obs_time}{Date and time.}
  \item{temp}{Air temperature (deg C).}
  \item{relhum}{Relative humidity (percent).}
  \item{pres}{Atmospheric pressure (kPa).}
  \item{swdown}{Downward shortwave radiation (W/m2).}
  \item{difrad}{Diffuse shortwave radiation (W/m2).}
  \item{lwdown}{Downward longwave radiation (W/m2).}
  \item{windspeed}{Wind speed (m/s).}
  \item{winddir}{Wind direction (degrees).}
  \item{precip}{Precipitation (mm).}
}
}
\usage{
climdata
}
\description{
Hourly weather observations for 2017 at Caerthillian Cove, Lizard, Cornwall
(49.96807 N, 5.215668 W).
}
\keyword{datasets}
