% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{clumpestimate}
\alias{clumpestimate}
\title{Estimate canopy clumpiness}
\usage{
clumpestimate(hgt, leafd, pai, maxclump = 0.95)
}
\arguments{
\item{hgt}{vegetation height (m).}

\item{leafd}{mean leaf width (m).}

\item{pai}{plant area index value(s).}

\item{maxclump}{optional parameter indicating maximum clumpiness}
}
\value{
`clump` parameter indicating the fraction of radiation passing directly through larger gaps in the canopy.
}
\description{
The function `clumpestimate` estimates how clumpy a canopy is based on height and
plant area index
}
\details{
`hgt`, `pai` and `leafd` must all be single numeric values or arrays with identical dimensions
}
