% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.biomicropoint}
\alias{.biomicropoint}
\title{Run micropoint for bioclim calculations}
\usage{
.biomicropoint(
  weather,
  precip,
  tme,
  vegp,
  soilc,
  reqhgt,
  windhgt,
  soilm,
  dTmx,
  maxiter,
  lat = NA,
  long = NA
)
}
\description{
Run micropoint for bioclim calculations
}
