% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.biosel}
\alias{.biosel}
\title{identifies which entries of weather time series should be selected}
\usage{
.biosel(tme, tc)
}
\description{
identifies which entries of weather time series should be selected
}
