% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.canopysnowint}
\alias{.canopysnowint}
\title{Calculates SWE (mm) of intercepted canopy snow (inputs as follows:)
prec: daily precipitation (mm),
uz: wind speed at height zu  (m/s),
gsnowd: ground snow depth (cm),
Lt: snow load in previous time step (mm SWE)
pai: plant area index
hgt: canopy height (m)
Sh: branch snow load coefficient (kg/m^2). 6.6 for pine and 5.9 kg for spruce.
zu: height above ground of wind speed measurement uz (m)}
\usage{
.canopysnowint(precd, uz, pai, hgt, gsnowd, Lt, Sh = 6.6, zu = 2)
}
\description{
Calculates SWE (mm) of intercepted canopy snow (inputs as follows:)
prec: daily precipitation (mm),
uz: wind speed at height zu  (m/s),
gsnowd: ground snow depth (cm),
Lt: snow load in previous time step (mm SWE)
pai: plant area index
hgt: canopy height (m)
Sh: branch snow load coefficient (kg/m^2). 6.6 for pine and 5.9 kg for spruce.
zu: height above ground of wind speed measurement uz (m)
}
