% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.cantransdif}
\alias{.cantransdif}
\title{Calculates diffuse radiation transmission through canopy}
\usage{
.cantransdif(l, ref = 0.23, clump = 0)
}
\description{
Calculates diffuse radiation transmission through canopy
}
