% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.cantransdir}
\alias{.cantransdir}
\title{Calculates direct radiation transmission through canopy}
\usage{
.cantransdir(l, k, ref = 0.23, clump = 0)
}
\description{
Calculates direct radiation transmission through canopy
}
