% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.cca}
\alias{.cca}
\title{Converts variable in a list of data.frame to an array as used by modelina}
\usage{
.cca(weather, varn, h, r, rfi)
}
\description{
Converts variable in a list of data.frame to an array as used by modelina
}
