% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.climds}
\alias{.climds}
\title{Selects relevent daily min (or max) data from hourly climate arrays}
\usage{
.climds(climarray, sel)
}
\description{
Selects relevent daily min (or max) data from hourly climate arrays
}
