% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.conductivityE}
\alias{.conductivityE}
\title{Calculates canopy boundary layer conductivity and conductivity to ground}
\usage{
.conductivityE(
  micro,
  reqhgt,
  xyf = NA,
  zf = NA,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA,
  maxhgt = NA
)
}
\description{
Calculates canopy boundary layer conductivity and conductivity to ground
}
