% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.dpsim}
\alias{.dpsim}
\title{Calculate integrated diabatic correction coefficient for momentum}
\usage{
.dpsim(ze)
}
\description{
Calculate integrated diabatic correction coefficient for momentum
}
