% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.epaif}
\alias{.epaif}
\title{Calculate pai above snow surface where snow depth is an array or matrix}
\usage{
.epaif(pai, hgt, snowd)
}
\description{
Calculate pai above snow surface where snow depth is an array or matrix
}
