% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.expandclim}
\alias{.expandclim}
\title{Expand daily climate variables to hourly (all)}
\usage{
.expandclim(mout_mn, mout_mx, micropoint, reqhgt)
}
\description{
Expand daily climate variables to hourly (all)
}
