% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowmodel.R
\name{.FitSnowOne}
\alias{.FitSnowOne}
\title{Fits snow model (one step)}
\usage{
.FitSnowOne(
  weather,
  snowdepth,
  precd,
  meltfact = 6.54,
  snowem = 0.99,
  zm = 0.002,
  umin = 0.5,
  astc = 1.5
)
}
\description{
Fits snow model (one step)
}
