% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.flowacc}
\alias{.flowacc}
\title{Caclulates flow accumulation (NB - probably needs modifying wiht below)}
\usage{
.flowacc(dtm)
}
\description{
Caclulates flow accumulation (NB - probably needs modifying wiht below)
}
