% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.gturb}
\alias{.gturb}
\title{Calculates turbulent molar conductivity above canopy}
\usage{
.gturb(uf, d, zm, z1, z0 = NA, psi_h = 0, gmin)
}
\description{
Calculates turbulent molar conductivity above canopy
}
