% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.hourtoday}
\alias{.hourtoday}
\title{Extract daily min, mean or max from array of hourly data}
\usage{
.hourtoday(a, fun = mean)
}
\description{
Extract daily min, mean or max from array of hourly data
}
