% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.is}
\alias{.is}
\title{Check if input is a SpatRaster or PackedSpatRaster and convert to matrix if it is}
\usage{
.is(r)
}
\description{
Check if input is a SpatRaster or PackedSpatRaster and convert to matrix if it is
}
