% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.latslonsfromr}
\alias{.latslonsfromr}
\title{Lats and longs from SpatRaster object, including reprojection}
\usage{
.latslonsfromr(r)
}
\description{
Lats and longs from SpatRaster object, including reprojection
}
