% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.meancanopywind}
\alias{.meancanopywind}
\title{Calculates wind speed within canopy given vertical profile within the canopy}
\usage{
.meancanopywind(uz, pai, hgt, d, zm, snowd = 0, zu = 2)
}
\description{
Calculates wind speed within canopy given vertical profile within the canopy
}
