% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.point0}
\alias{.point0}
\title{Calculate point surface temperature without diabatic coefficients and G set to 0}
\usage{
.point0(micropoint)
}
\description{
Calculate point surface temperature without diabatic coefficients and G set to 0
}
