% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.prepsnowinputs1}
\alias{.prepsnowinputs1}
\title{prepare inputs for snow microclimate model (data.frame climate)}
\usage{
.prepsnowinputs1(
  reqhgt,
  dtm,
  vegp,
  soilc,
  micropoints,
  snowdays,
  nosnowdays,
  smod,
  moutn,
  tmeorig,
  subs,
  runchecks,
  slr = NA,
  apr = NA,
  hor = NA,
  svf = NA,
  wsa = NA,
  paia
)
}
\description{
prepare inputs for snow microclimate model (data.frame climate)
}
