% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.prepsnowinputs2}
\alias{.prepsnowinputs2}
\title{prepare inputs for snow microclimate model (array climate)}
\usage{
.prepsnowinputs2(
  reqhgt,
  dtm,
  dtmc,
  vegp,
  soilc,
  micropoints,
  altcorrect,
  runchecks,
  snowdays,
  nosnowdays,
  moutn,
  slr = NA,
  apr = NA,
  hor = NA,
  svf = NA,
  wsa = NA,
  paia
)
}
\description{
prepare inputs for snow microclimate model (array climate)
}
