% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.psunlitd}
\alias{.psunlitd}
\title{Calculates equivelent of proportion of sunlight leaves for diffuse radiation}
\usage{
.psunlitd(l, clump = 0)
}
\description{
Calculates equivelent of proportion of sunlight leaves for diffuse radiation
}
