% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.rhcanopy}
\alias{.rhcanopy}
\title{Calculate inverse thermal diffusivity below canopy}
\usage{
.rhcanopy(uf, h, d, z, phih = 1)
}
\description{
Calculate inverse thermal diffusivity below canopy
}
