% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.roughlength}
\alias{.roughlength}
\title{Calculate roughness length governing momentum transfer}
\usage{
.roughlength(hgt, pai, zm0 = 0.003)
}
\description{
Calculate roughness length governing momentum transfer
}
