% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.roughtlength}
\alias{.roughtlength}
\title{Calculate roughness length governing momentum transfer}
\usage{
.roughtlength(h, pai)
}
\description{
Calculate roughness length governing momentum transfer
}
