% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.runmicronosnow}
\alias{.runmicronosnow}
\title{Runs microclimate model without snow}
\usage{
.runmicronosnow(
  micropoint,
  reqhgt,
  vegp,
  soilc,
  dtm,
  dtmc = NA,
  altcorrect = 0,
  runchecks = TRUE,
  pai_a = NA,
  tfact = 1.5,
  out = rep(TRUE, 10),
  slr = NA,
  apr = NA,
  hor = NA,
  twi = NA,
  wsa = NA,
  svf = NA,
  method = "Cpp"
)
}
\description{
Runs microclimate model without snow
}
