% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.runmicrosnow1}
\alias{.runmicrosnow1}
\title{Run microclimate grid model with snow (data.frame weather)}
\usage{
.runmicrosnow1(
  micropoint,
  reqhgt,
  vegp,
  soilc,
  dtm,
  smod,
  runchecks = TRUE,
  pai_a = NA,
  tfact = 1.5,
  out = rep(TRUE, 10),
  slr = NA,
  apr = NA,
  hor = NA,
  twi = NA,
  wsa = NA,
  svf = NA
)
}
\description{
Run microclimate grid model with snow (data.frame weather)
}
