% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.runmicrosnow2}
\alias{.runmicrosnow2}
\title{Run microclimate grid model with snow (array weather)}
\usage{
.runmicrosnow2(
  micropoint,
  reqhgt,
  vegp,
  soilc,
  dtm,
  dtmc,
  smod,
  altcorrect = 0,
  runchecks = TRUE,
  pai_a = NA,
  tfact = 1.5,
  out = rep(TRUE, 10),
  slr = NA,
  apr = NA,
  hor = NA,
  twi = NA,
  wsa = NA,
  svf = NA
)
}
\description{
Run microclimate grid model with snow (array weather)
}
