% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.runmodel1Cpp}
\alias{.runmodel1Cpp}
\title{Run C++ version of model with time-invarient vegetation and data.frame weather input}
\usage{
.runmodel1Cpp(
  micropoint,
  vegp,
  soilc,
  dtm,
  reqhgt = 0.05,
  runchecks = TRUE,
  pai_a = NA,
  tfact = 1.5,
  out = rep(TRUE, 10),
  slr = NA,
  apr = NA,
  hor = NA,
  twi = NA,
  wsa = NA,
  svf = NA
)
}
\description{
Run C++ version of model with time-invarient vegetation and data.frame weather input
}
