% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.runmodel2Cpp}
\alias{.runmodel2Cpp}
\title{Run C++ version of model with time-invarient vegetation and array weather input}
\usage{
.runmodel2Cpp(
  micropointa,
  vegp,
  soilc,
  dtm,
  dtmc,
  reqhgt = 0.05,
  runchecks = TRUE,
  altcorrect = 0,
  pai_a = NA,
  tfact = 1.5,
  out = rep(TRUE, 10),
  slr = NA,
  apr = NA,
  hor = NA,
  twi = NA,
  wsa = NA,
  svf = NA
)
}
\description{
Run C++ version of model with time-invarient vegetation and array weather input
}
