% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.selquarter}
\alias{.selquarter}
\title{identifies which entries of bioclim time series should be selected for a given quarter}
\usage{
.selquarter(tmeh, mon)
}
\description{
identifies which entries of bioclim time series should be selected for a given quarter
}
