% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.snowdens}
\alias{.snowdens}
\title{Calculates snow density parameters from snow environment}
\usage{
.snowdens(snowenv = "Tundra")
}
\description{
Calculates snow density parameters from snow environment
}
