% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowmodel.R
\name{.SnowEnergyBalance}
\alias{.SnowEnergyBalance}
\title{Computes snow energy balance}
\usage{
.SnowEnergyBalance(
  tc,
  u2,
  rh,
  pk,
  Rsw,
  skyem,
  snowalb,
  snowtemp,
  snowem = 0.99,
  zm = 0.002,
  umin = 0.5
)
}
\description{
Computes snow energy balance
}
