% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.snowmodel1}
\alias{.snowmodel1}
\title{full snow model with data.frame weather inputs}
\usage{
.snowmodel1(
  weather,
  dtm,
  vegp,
  soilc,
  snowenv = "Taiga",
  snowinitd = 0,
  snowinita = 0,
  zref = 2,
  windhgt = zref,
  tfact = 0.02
)
}
\description{
full snow model with data.frame weather inputs
}
