% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.snowmodel2}
\alias{.snowmodel2}
\title{full snow model with array weather inputs}
\usage{
.snowmodel2(
  weather,
  tme,
  dtm,
  dtmc,
  vegp,
  soilc,
  altcorrect = 0,
  snowenv = "Taiga",
  snowinitd = 0,
  snowinita = 0,
  zref = 2,
  windhgt = zref,
  tfact = 0.02
)
}
\description{
full snow model with array weather inputs
}
