% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.snowmodelq1}
\alias{.snowmodelq1}
\title{quick snow model with data.frame weather inputs}
\usage{
.snowmodelq1(
  weather,
  dtm,
  vegp,
  soilc,
  subs,
  snowenv = "Taiga",
  snowinitd = 0,
  snowinita = 0,
  zref = 2,
  windhgt = zref,
  tfact = 0.02
)
}
\description{
quick snow model with data.frame weather inputs
}
