% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.snowmodelq2}
\alias{.snowmodelq2}
\title{quick snow model with array weather inputs}
\usage{
.snowmodelq2(
  weather,
  tme,
  dtm,
  dtmc,
  vegp,
  soilc,
  subs,
  altcorrect = 0,
  snowenv = "Taiga",
  snowinitd = 0,
  snowinita = 0,
  zref = 2,
  windhgt = zref,
  tfact = 0.02
)
}
\description{
quick snow model with array weather inputs
}
