% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.solvemicro}
\alias{.solvemicro}
\title{Iteratively uns a point microclimate model for heat exchange surface of canopy}
\usage{
.solvemicro(
  micro,
  climdata,
  d,
  zm,
  reqhgt,
  backweight = 2,
  maxiter = 100,
  gmn = 0.3,
  ti
)
}
\description{
Iteratively uns a point microclimate model for heat exchange surface of canopy
}
