% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.sortl2}
\alias{.sortl2}
\title{calculate average vegetation value for hours with snow with paia and foliage density}
\usage{
.sortl2(vegp, sdep, reqhgt, pai_a)
}
\description{
calculate average vegetation value for hours with snow with paia and foliage density
}
