% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.sortrough}
\alias{.sortrough}
\title{Sort out smoothing of zero-plane displacement height and roughness length}
\usage{
.sortrough(vegp, xyf)
}
\description{
Sort out smoothing of zero-plane displacement height and roughness length
}
