% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.subsetpointtoday}
\alias{.subsetpointtoday}
\title{Function for selecting daily min and max from point model}
\usage{
.subsetpointtoday(micropoint)
}
\description{
Function for selecting daily min and max from point model
}
