% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.t0f}
\alias{.t0f}
\title{Calculates reference time for phase of diurnal temperature fluctuation}
\usage{
.t0f(tc)
}
\description{
Calculates reference time for phase of diurnal temperature fluctuation
}
