% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.t0fd}
\alias{.t0fd}
\title{Calculates reference time for phase of diurnal temperature fluctuation (daily)}
\usage{
.t0fd(tc)
}
\description{
Calculates reference time for phase of diurnal temperature fluctuation (daily)
}
