% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.TGadjust}
\alias{.TGadjust}
\title{Adjust canopy temperature by gorund temperature}
\usage{
.TGadjust(dT, gHr, m, delta, tc, T0, gCa)
}
\description{
Adjust canopy temperature by gorund temperature
}
