% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.tleaf}
\alias{.tleaf}
\title{Calculate temperature profile below canopy}
\usage{
.tleaf(
  tair,
  tground,
  es,
  eref,
  pk,
  gtt,
  gt0,
  gHa,
  gv,
  gL,
  Rabsl,
  leafdens,
  surfwet,
  soilrh = 1
)
}
\description{
Calculate temperature profile below canopy
}
