% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.todf}
\alias{.todf}
\title{Converts array of climate data to a dataframe for a given cell}
\usage{
.todf(climarray, i, j, tme, winddir)
}
\description{
Converts array of climate data to a dataframe for a given cell
}
