% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.windcoef}
\alias{.windcoef}
\title{Calculates wind shelter coefficient in specified direction}
\usage{
.windcoef(dsm, direction, hgt = 1, res = 1)
}
\description{
Calculates wind shelter coefficient in specified direction
}
