% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.windshelter}
\alias{.windshelter}
\title{Calculate wind speeds accounting for topographic sheltering}
\usage{
.windshelter(
  u2,
  wdir,
  dsm,
  whgt = 2,
  wdct = 8,
  xyf = 10,
  hs = TRUE,
  wsa = NA,
  cors = NA
)
}
\description{
Calculate wind speeds accounting for topographic sheltering
}
