% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.windsheltera}
\alias{.windsheltera}
\title{Calculates array of wind shelter coefficients in wdct directions}
\usage{
.windsheltera(dsm, wdct = 8, whgt = 2, xyf = 10, hs = TRUE)
}
\description{
Calculates array of wind shelter coefficients in wdct directions
}
