% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtmcaerth}
\alias{dtmcaerth}
\title{Digital terrain model for Caerthillian Cove}
\format{
A 50 x 50 cell \code{PackedSpatRaster}.
}
\source{
\url{http://www.tellusgb.ac.uk/}
}
\usage{
dtmcaerth
}
\description{
One-metre-resolution elevation (m) for a 50 x 50 m area of Caerthillian
Cove, Cornwall, on the British National Grid (EPSG:27700; extent 169475--169525
E, 12475--12525 N).
}
\keyword{datasets}
