% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{expandtonc}
\alias{expandtonc}
\title{expand runmicro_big daily output to hourly and write as ncff4 file}
\usage{
expandtonc(filein, fileout)
}
\arguments{
\item{filein}{file to be loaded as written out by [runmicro_big()]}

\item{file}{to be written out. Must have .nc extension.}
}
\description{
the function `expandtonc` takes the daily output written by [runmicro_big()]
when hourly = F, expands this to hourly, and writes data out as an ncdf4 file.
}
