% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowmodel.R
\name{fitsnow}
\alias{fitsnow}
\title{Derives snow melt coefficient from empirical data}
\usage{
fitsnow(
  weather,
  snowdepth,
  precd,
  plotout = T,
  snowem = 0.99,
  zm = 0.002,
  umin = 0.5,
  astc = 1.5
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details).}

\item{precd}{a vector of daily precipitation (mm).}

\item{plotout}{optional logical indicating whether to plot observed and predicted data}

\item{snowem}{optionally, numeric value of snow emissivity}

\item{zm}{optionally, numeric value of roughness length for momentum transfer of snow surface (m)}

\item{umin}{optionally, numeric value indicating minimum wind speed used in conductivity calculations (m/s)}

\item{astc}{optionally, numeric value indicating the temperature at which precipitation falls as snow (deg C)}

\item{vector}{of empirically-derived hourly snow depths (cm)}
}
\value{
a list of the following:

`psnowdepth` predicted snow depth (cm)

`RMS` Root-mean-square error of predicted snow depths (cm)

`meltfact` a snow melt factor for use with [pSnow()]
}
\description{
The function `fitsnow` estimates the snow melt coefficient from empirical data (or
the outputs of a more complex model) and returns this along with
estimates of snow depth and the RMS error
}
\details{
The format and and units of `weather` must follow that in the example
dataset `climdata`.
}
