% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilfitfunctions.R
\name{fitsoilm}
\alias{fitsoilm}
\title{Function for deriving parameters for soil moisture model}
\usage{
fitsoilm(weather, rainfall, soilm, soiltype, plotout = TRUE, max.iter = 20)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details).}

\item{rainfall}{a vector of daily rainfall.}

\item{soilm}{a vector of daily soil moisture values.}

\item{soiltype}{a soil type as listed in [soilparameters()].}

\item{plotout}{optional logical indicating whether to plot fit on eahc iteration.}

\item{max.iter}{maximum number of iterations used for fitting the model.}
}
\value{
a list wiht the following components:

`mult` multiplier for net radiation.

`rmu`  multiplier for rainfall.

`a` constant controlling sub-surface and surface water exchange.

`pwr` exponent controlling sub-surface and surface water exchange.

`rms` RMS Error of model fit.
}
\description{
The function `fitsoilm` derives parameters for fitting the soil moisture model
where daily moisture data are available
}
\details{
The format and and units of `weather` must follow that in the example
dataset `climdata`. In this version of the model, all precipitation is assumed to
fall as rain.
}
\examples{
require(NicheMapR) # - see https://github.com/mrke/NicheMapR
require(microclimc) # - see https://github.com/ilyamaclean/microclimc
# Run NicheMapR complex multilayer soil model
microout<-runNMR(climdata, rainfall, 50.2178, -5.32656, 0.05, 0.01, PAI = 0, soiltype="Clay loam")
# Extract surface soil moisture and covert to daily
sm<-microout$soilmoist
soilm<-sm[,3]
soilm<-matrix(soilm,ncol=24,byrow=T)
soilm<-apply(soilm,1,mean)
# Fit simple model
soilmcoefs<-fitsoilm(weather, rainfall, soilm, "Clay")
soilmcoefs
}
