% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilfitfunctions.R
\name{fitsoilt}
\alias{fitsoilt}
\title{Function for deriving parameters for soil temperature model}
\usage{
fitsoilt(
  weather,
  rainfall,
  soiltemp,
  soiltype,
  theta = NA,
  plotout = TRUE,
  soilmcoefs = NA
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details).}

\item{rainfall}{a vector of daily rainfall.}

\item{soiltemp}{a vector of hourly soil temperatures}

\item{soiltype}{a soil type as listed in [soilparameters()].}

\item{theta}{optional vector of hourly soil moistures. If NA estimated using [soilmpredict()]}

\item{plotout}{optional logical indicating whether to plot fit on eahc iteration.}

\item{soilmcoefs}{an optional list of soil moisture model coefficients as returned by [fitsoilm()] (see details)}
}
\value{
a list with the following components:

`int` model intercept

`t1`  multiplier for net radiation.

`t2` multiplier for soil moisture.

`t3` multiplier for interaction term.

`rms` RMS Error of model fit.
}
\description{
The function `fitsoilt` derives parameters for fitting the soil temperature model
where hourly temperature data are available
}
\details{
The format and and units of `weather` must follow that in the example
dataset `climdata`. If `soilmcoefs` is not supplied, the inbuilt parameters in
[soilparameters()] are used to derive soil moisture.
}
\examples{
require(NicheMapR) # - see https://github.com/mrke/NicheMapR
require(microclimc) # - see https://github.com/ilyamaclean/microclimc
# Run NicheMapR complex multilayer soil model
microout<-runNMR(climdata, rainfall, 50.2178, -5.32656, 0.05, 0.01, PAI = 0, soiltype="Clay loam")
# Extract surface soil temperature
st<-microout$soiltemps
soiltemp<-st$D0cm
# Fit simple model
soiltcoefs<-fitsoilt(climdata, rainfall, soiltemp, "Clay loam")
soiltcoefs
}
