% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{foliageden}
\alias{foliageden}
\title{Estimate foliage density and plant area index above z}
\usage{
foliageden(z, hgt, pai, shape = 1.5, rate = shape/7)
}
\arguments{
\item{z}{height above ground (m)}

\item{hgt}{the height of the canopy (m)}

\item{pai}{total plant area index of canopy}

\item{shape}{optional shape parameter for gamma distribution (see details)}

\item{rate}{optional rate parameter for gamm distribution (see details)}
}
\value{
a list with the following components:

`leafden` the foliage density (one sided leaf area per m^3) at height z

`paia` the plant area index above `z`

a list of leaf densities and plant area index values
}
\description{
The function `foliageden` applies a mirrored gamma distribution
to estimate, for a given height `z` below canopy, the foliage density at
height z and the plant area index above `z`
}
\details{
a broadly realistic estimate of foliage density is generated
using a mirrored gamma distribution (i.e. right (top) rather than left-skewed).
in applying the gamma distribution `z / hgt` is re-scaled to the range 0-10.
}
\examples{
z<-c(1:1000)/100
fdp<-foliageden(z,10,10)
plot(z~fdp$leafden, xlab = "Foliage density")
plot(z~fdp$pai_a, xlab = "PAI above z")
}
