% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globclim}
\alias{globclim}
\title{Global climate summary data}
\format{
An array with 94 rows, 192 columns and five climate variables:
\describe{
  \item{1}{Mean annual temperature (deg C).}
  \item{2}{Coefficient of variation in temperature (K).}
  \item{3}{Mean annual rainfall (mm per year).}
  \item{4}{Coefficient of variation in annual rainfall (mm per 0.25 days).}
  \item{5}{Calendar month with the greatest rainfall (1--12).}
}
}
\source{
\url{http://www.ncep.noaa.gov/}
}
\usage{
globclim
}
\description{
Global climate summaries averaged over 2008--2017 on a 94 x 192 grid.
}
\keyword{datasets}
