% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{groundrad}
\alias{groundrad}
\title{Calculates radiation absorbed by ground}
\usage{
groundrad(micro, slr = NA, apr = NA, hor = NA)
}
\arguments{
\item{micro}{an object of class `micro` as returned by [modelin()]}

\item{slr}{an optional SpatRaster object of slope values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional SpatRaster object of aspect values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied, but outer cells will be NA.}
}
\value{
an object of class micro with the following components added:

`radGsw` shortwave radiation absorbed by the ground (W/m^2)

`radGlw` longwave radiation absorbed by the ground (W/m^2)

additional terms needed for subsequent modelling
}
\description{
The function `groundrad` calculates shortwave and longwave
radiation absorbed by the ground
}
