% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimf-package.R
\docType{package}
\name{microclimf-package}
\alias{microclimf}
\alias{microclimf-package}
\title{microclimf: Fast above, below or within canopy gridded microclimate modelling with R}
\description{
The functions in the microclimf package contain tools for fast modelling of the mechanistic processes that govern below or above canopy microclimate on gridded data. It builds on the microclima and microclimc packages, but contains all the functions needed to run the model in isolation, with no other microclimate package dependancies. A simplified means of estimating soil temperatures is included, eliminating the need to run the model in time-steps. In consequences calculations are performed simultaniously on all grid cells and over all time-increments. The models can be run either in hourly or in daily time-increments, the latter quicker with with lower memory requirements. In interpolation method to derive hourly estimates from daily data is also included. This package is a piecemeal rebuild of microclimf; at this stage it contains only the package datasets, with functions to be reintroduced incrementally.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ilyamaclean/microclimf}
  \item Report bugs at \url{https://github.com/ilyamaclean/microclimf/issues}
}

}
\author{
\strong{Maintainer}: Ilya Maclean \email{i.m.d.maclean@exeter.ac.uk}

}
\keyword{internal}
