% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{nctoclimarray}
\alias{nctoclimarray}
\title{Create climate arrays for inputting to microclimate model from a netCDF4 file}
\usage{
nctoclimarray(ncfile, dtm, dtr_cor_fac = 1.285)
}
\arguments{
\item{ncfile}{character vector containing the path and filename of the nc file}

\item{dtm}{a SpatRaster object of elevations covering the extent of the study area (see details)}

\item{dtr_cor_fac}{numeric value to be used in the diurnal temperature range
correction of coastal grid cells. Default = 1.285, based on calibration against UK Met Office
observations. If set to zero, no correction is applied.}
}
\value{
a list of the following:
\describe{
  \item{tme}{POSIXlt object of times corresponding to climate observations}
  \item{climarray}{a list of arrays of hourly weather variables - same format as for [modelina()].}
  \item{precarray}{an array of daily precipitation values - same format as for [modelina()].}
  \item{dtmc}{a coarse resolution digital elevation dataset matching the resolution of input
  climate data, but with a coordinate reference system and extent matching `dtm`}
}
}
\description{
The function `nctoclimarray` converts data in a netCDF4 file returned
by [mcera5::request_era5()] to the correct formal required for running [modelina()] or
[modelina_dy()].
}
\details{
the model requires that input climate data are projected using a coordinate reference
system in which x and y are in metres. Since values returned by [mcera5::request_era5()]
are in lat long, the output data are reprojected using the coordinate reference system and
extent of dtm (but retain the approximate original grid resolution of the input climate data).
Returned climate data match the resolution, corrdinate reference system and extent of `dtmc`.
}
