% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{PenMont}
\alias{PenMont}
\title{Applies Penman-Monteith equation}
\usage{
PenMont(
  tc,
  pk,
  ea,
  radabs,
  gHa,
  gV,
  es = NA,
  tdew = NA,
  surfwet = 1,
  G = 0,
  allout = TRUE
)
}
\arguments{
\item{tc}{air temperature (deg c)}

\item{pk}{atmospheric pressure (kPa)}

\item{ea}{vapour pressure of air (kPa)}

\item{radabs}{absorbed long and shortwave radiation (W/m2)}

\item{gHa}{boundary layer conductances for heat (mol/m^2/s)}

\item{gV}{conductance for vapour (mol/m^2/s)}

\item{es}{optional saturated vapour pressure (kPa) at temperature tc (calculated if NA)}

\item{tdew}{optional dewpoint temperature (deg C) (calculated if NA, see details)}

\item{surfwet}{proportion of surface acting as wet surface}

\item{G}{optional ground heat flux (W/m^2)}

\item{allout}{optional logical indicating whether to return Latent, Sensible and
Emitted radiation flux in addition to surface temperature}
}
\value{
a list with the following components:

`T0` surface temperature (deg C)

`H` the sensible heat flux (w/m^2) (NA if allout set to FALSE)

`L` the latent heat flux (w/m^2) (NA if allout set to FALSE)

`Rem` the emitted radiation flux (w/m^2) (NA if allout set to FALSE)
}
\description{
The function `PenMont` applies a varient of the Penman-Monteith
equation, to calculate surface temperature and the energy fluxes at the surface
}
\details{
The paramater `surfwet` determines how much of the canopy should be treated as
wet surface. However, except when extremely droughted, the matric potential of
of soil leaves is such that `surfwet` ~ 1.
}
