% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowmodel.R
\name{pSnow}
\alias{pSnow}
\title{Estimates snow depth and other snow variables}
\usage{
pSnow(
  weather,
  precd,
  meltfact = 6.54,
  snowem = 0.99,
  zm = 0.002,
  umin = 0.5,
  astc = 1.5
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details).}

\item{precd}{a vector of daily precipitation (mm).}

\item{meltfact}{a snow melt coefficient, as returned by [FitSnow()]}

\item{snowem}{optionally, numeric value of snow emissivity}

\item{zm}{optionally, numeric value of roughness length for momentum transfer of snow surface (m)}

\item{umin}{optionally, numeric value indicating minimum wind speed used in conductivity calculations (m/s)}

\item{astc}{optionally, numeric value indicating the temperature at which precipitation falls as snow (deg C)}
}
\value{
a list of the following:

`snowdepth` snow depth (cm)

`snowdens` snow density (g/cm^3)

`snowtemp` snow temperature (deg C)

`snowalb` snow albedo (range 0 to 1)
}
\description{
The function `pSnow` estimates snow depth, albedo, density and temperature
}
\details{
The format and and units of `weather` must follow that in the example
dataset `climdata`.
}
