% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Daily rainfall}
\format{
A vector of daily rainfall (mm/day)
}
\usage{
rainfall
}
\description{
A vector of daily rainfall for 169500, 12500 (x, y) using the Ordance Survey GB Grid Reference system (CRS: 27700).
}
\keyword{datasets}
