% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{resampleclimdata}
\alias{resampleclimdata}
\title{Resamples an a list of SpatRaster climate variables to an appropriate resolution
for microclimate modelling}
\usage{
resampleclimdata(climarrayr, dtm)
}
\arguments{
\item{climarrayr}{a list of climate variables as SpatRasters or PackedSpatRasters}

\item{a}{SpatRaster or PackedSpatRaster of digital elevations used to delineate the study area}
}
\description{
if the resolution of `climarrayr` is coarse and covers a wider extent than
than the study area determined by `dtm`, the climate array version of the model way not
work. This function allows climate data to be cropped and resampled to an appropriate
resolution and extent for running the array version of the model
}
