% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimfunctions.R
\name{runbioclim_big}
\alias{runbioclim_big}
\title{The function `runbioclim_big` runs the microclimate model and produces microclimate
equivalents of the 19 Worldclim bioclimate variables in tiles over large areas}
\usage{
runbioclim_big(
  weather,
  precip,
  tme = NA,
  reqhgt = 0.05,
  vegp,
  soilc,
  dtm,
  dtmc = NA,
  temp = "air",
  hourly = TRUE,
  altcorrect = 1,
  windhgt = 2,
  soilm = NA,
  dTmx = 25,
  maxiter = 100,
  runchecks = TRUE,
  xyf = 1,
  pai_a = NA,
  tfact = 1.5,
  surfwet = NA,
  silent = FALSE
)
}
\arguments{
\item{weather}{a data.frame or list of arrays of weather variables (as for [runpointmodel() or [runpointmodela()])}

\item{precip}{a vector or array of daily precipitation variables ((as for [runpointmodel() or [runpointmodela()]))}

\item{tme}{if weather is not a data.frame, an object of class POSIXlt giving the dates and times for each weather variable stored in the array. Set to NA if weather is a data,frame}

\item{reqhgt}{height for which temperatures are needed}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()]}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{dtm}{a SpatRaster object of elevations in metres}

\item{dtmc}{If weather is s not a data.frame, a SpatRaster object giving the resolution,
spatial extent, and projection of the weather data. Must give elevations in meters if
`altcorrect` > 0.}

\item{temp}{one of "air" or "leaf" indicating whether bioclim variables are constructed using air or leaf
temperatures. If `temp = leaf`, for grid cells where `reqhgt` is above vegetation, vertically averaged canopy
temperature is used.}

\item{hourly}{optional logical indicating whether to run the model in hourly or daily mode.}

\item{altcorrect}{a single numeric value indicating whether to apply an elevational lapse rate correction to temperatures (0 = no correction, 1 = fixed lapse rate correction, 2 = humidity-dependent variable lapse rate correction)}

\item{windhgt}{height above ground of wind speed data in weather}

\item{soilm}{optional vector of soil moisture values in upper 10 cm of the soil (calculated if not supplied)}

\item{dTmx}{optional maximum amount by which canopy or ground surface temperatures can exceed air temperatures.
Included to ensure model convergence.}

\item{maxiter}{optional integer indicating the maximum number of iterations (see details).}

\item{runchecks}{optional logical indicating whether to call [checkinputs()] to run}

\item{xyf}{optional spatial smoothing factor applied in calculation of surface
roughness and zero-plane displacement heights (see [modelin()])}

\item{pai_a}{an optional array of plant area index values above `reqhgt`.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation in topographic wetness (see [soilmdistribute()]).}

\item{surfwet}{an optional single numeric value of array of values specifying the proportion
of the canopy surface that should be treated as wet surface (modelled if not supplied)}

\item{silent}{optional logical indicating whether to report on progress (default FALSE - progress reported).}
}
\value{
a multilayer SpatRast of the following:
\describe{
 \item{BIO1}{Mean of monthy median temperatures (degrees C)}
 \item{BIO2}{Mean diurnal temperature range (degrees C)}
 \item{BIO3}{Isothermality (BIO2/BIO7) (×100)}
 \item{BIO4}{Temperature Seasonality (standard deviation of monthly median temperatures × 100)}
 \item{BIO5}{Maximum temperature (degrees C)}
 \item{BIO6}{Minimum temperature (degrees C)}
 \item{BIO7}{Temperature Annual Range (BIO5-BIO6) (degrees C)}
 \item{BIO8}{Mean of monthly median temperatures in wettest three months (degrees C)}
 \item{BIO9}{Mean of monthly median temperatures in driest three months (degrees C)}
 \item{BIO10}{Mean of monthly median temperatures in warmest three months (degrees C)}
 \item{BIO11}{Mean of monthly median temperatures in coldest three months (degrees C)}
 \item{BIO12}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of each month (m^3 / m^3)}
 \item{BIO13}{Maximum of monthly soil moistures on the hottest, coldest and median temperature day of each month (m^3 / m^3)}
 \item{BIO14}{Minimum of monthly soil moistures on the hottest, coldest and median temperature day of each month (m^3 / m^3)}
 \item{BIO15}{Soil moisture seasonality (Coefficient of Variation) on day in each month with median temperature}
 \item{BIO16}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of the wettest three months (m^3 / m^3)}
 \item{BIO17}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of the driest three months (m^3 / m^3)}
 \item{BIO18}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of the warmest three months (m^3 / m^3)}
 \item{BIO19}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of the coldest three months (m^3 / m^3)}
}
}
\description{
The function `runbioclim_big` runs the microclimate model and produces microclimate
equivalents of the 19 Worldclim bioclimate variables in tiles over large areas
}
\details{
The model is run in 100 pixel by 100 pixel tiles and the tiles are automatically merged back-togther
to create a single merged output. When climate data are provided as arrays tiling effects can be
present, and to avoid this, the tiles are run with a 10 pixel overlap on each side, and distance-weighted
blending undertaken to derive the final output.
}
\seealso{
[runbioclim()]
}
