% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonhourly.R
\name{runpointmodela}
\alias{runpointmodela}
\title{Runs point microclimate model over arrays}
\usage{
runpointmodela(
  climarray,
  precarray,
  tme,
  vegp,
  soilc,
  windhgt = 2,
  reqhgt = 0.1,
  soilm = NA,
  dTmx = 25,
  maxiter = 10
)
}
\arguments{
\item{climarray}{a list of arrays of weather variables (see details and [nctoclimarray()])}

\item{precarray}{an array of daily precipitation (see details)}

\item{tme}{an object of class POSIXlt giving the dates and times for each weather variable stroed in the array}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()] (see details)}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{windhgt}{height above ground of wind speed data in weather}

\item{reqhgt}{height for which temperatures are needed (used only when reqhgt < 0 to calculate tmeperature below ground)}

\item{soilm}{optional vector of soil moisture values in upper 10 cm of the soil (calculated if not supplied)}

\item{dTmx}{optional maximum amount by which canopy or ground surface temperatures can exceed air temperatures.
Included to ensure model convergence}

\item{maxiter}{optional integer indicating the maximum number of iterations (see details)}
}
\value{
a list of point model outputs (as returned by [runpointmodel()] for each grid cell
}
\description{
The function `runpointmodela` runs the point microclimate model over every cell of an array of climate data
}
\details{
The units of `climarray` must follow those in the dataset `climdata`.
It must be a list with each component of the list an array, named using the same
names as the column headers in weather (e.g. temp for temperature), excluding `obs_time`.
As not all wind measurements are at reference height, the height of the wind speed measurement
must be specified if not 2 m. To enable calculation of below-canopy wind and temperature profiles
in tall canopy, the wind speed and temperature data are  adjusted to give values for a height at
the maximum vegetation height if the tallest vegetation exceeds two metres.
For doing so a stand vegetation surface typical of that in which a weather station would be located is assumed.
The parameter `maxiter` sets the maximum number of times the model is iterated to achieve
convergence. Increasing this value improves accuracy at the expense of computation time.
}
