% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snowenv}
\alias{snowenv}
\title{Global snow-environment classes}
\format{
A 72 x 144 cell \code{PackedSpatRaster}.
}
\source{
Derived from \url{https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world}.
}
\usage{
snowenv
}
\description{
A 2.5-degree global classification of snow environments: 0 = Alpine,
1 = maritime, 2 = Prairie, 3 = Taiga, and 4 = Tundra.
}
\keyword{datasets}
