% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilc}
\alias{soilc}
\title{Soil properties at Caerthillian Cove}
\format{
A list with:
\describe{
  \item{soiltype}{Integer soil-class \code{PackedSpatRaster}; codes correspond to \code{soilparamstable$Number}.}
  \item{groundr}{Ground/litter shortwave reflectance \code{PackedSpatRaster}.}
}
}
\usage{
soilc
}
\description{
Spatial soil class and ground shortwave reflectance for the Caerthillian Cove
example landscape.
}
\keyword{datasets}
