% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{soilmpredict}
\alias{soilmpredict}
\title{Predicts soil moisture from rainfall and radiation}
\usage{
soilmpredict(rainfall, rnet, soiltype, soilinit = c(NA, NA), soilmcoefs = NA)
}
\arguments{
\item{rainfall}{a vector of daily rainfall (mm/day)}

\item{rnet}{a vector of hourly net radiation (W/m^2)}

\item{soiltype}{a soil type (as listed in [soilparameters()])}

\item{soilinit}{initial soil moisture fractions in surface and sub-surface layers}

\item{soilmcoefs}{an optional list of coefficients for running the soil moisture model
as returned by [fitsoilm()]. Uses inbuilt parameters if not supplied.}
}
\value{
a list of the following:

`soilm1` soil moisture fraction in surface layer

`soilm2` soil moisture fraction in sub-surface layer
}
\description{
The function `soilmpredict` implements a simple two layer
soil model to predict the daily soil moisture fraction in the surface and
sub-surface layers
}
\examples{
# Calculate net radiation from inbuilt climate dataset
swrad <- (1 - 0.15) * climdata$swrad
lwout <- 5.67e-8 * 0.95 * (climdata$temp + 273.15)^4
lwnet <- (1 - climdata$skyem) * lwout
rnet <- swrad - lwnet
# Run soil moisture model using inbuilt rainfall dataset
sm<-soilmpredict(rainfall, rnet, "Loam", c(0.4, 0.4))
# Plot results
plot(sm$soilm1, type="l", ylim=c(0.1, 0.5), xlab="Day",
     ylab="Soil moisture fraction")
par(new = TRUE)
plot(sm$soilm2, type="l", ylim=c(0.1,0.5), col="blue", lwd=2,
xlab="", ylab="")
}
