% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilparameters}
\alias{soilparameters}
\title{A table of soil parameters}
\format{
A data frame with the following columns:
\describe{
  \item{Soil.type}{description of soil type}
  \item{Number}{an integer value corresponding to each soil type}
  \item{Smax}{Volumetric water content at saturation (m^3 / m^3)}
  \item{Smin}{Residual water content (m^3 / m^3)}
  \item{Ksat}{Saturated hydraulic conductivity (kg / m^3 / day)}
  \item{b}{Shape parameter for Campbell soil moisture model (dimensionless, > 1)}
  \item{psi_e}{Matric potential (J / m^3)}
  \item{Vq}{Volumetric quartz content of soil}
  \item{Vm}{Volumetric mineral content of soil}
  \item{Vo}{Volumetric organic content of soil}
  \item{Mc}{Mass fraction of clay}
  \item{rho}{Soil bulk density (Mg / m^3)}
  \item{int}{temperature model intercept}
  \item{t1}{temperature model radiation coefficient}
  \item{t2}{temperature model soil moisture coefficient}
  \item{t3}{temperature model wind coefficient}
  \item{t4}{temperature model soil moisture * radiation interaction coefficient}
  \item{t5}{temperature model soil moisture * wind interaction coefficient}
  \item{t6}{temperature model radiation * wind interaction coefficient}
  \item{t7}{temperature model three-way interaction coefficient}
  \item{mult}{soil moisture model radiation coefficient}
  \item{rmu}{soil moisture model rainfall coefficient}
  \item{a}{soil moisture model deeper layer multiplier coefficient}
  \item{pwr}{soil moisture model deeper layer power coefficient}
}
}
\source{
: \url{https://onlinelibrary.wiley.com/doi/full/10.1002/ird.1751}
}
\usage{
soilparameters
}
\description{
A table of soil parameters for different soil types
}
\keyword{datasets}
