% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilparamsp}
\alias{soilparamsp}
\title{A dataset of vegetation parameters.}
\format{
A list of  the following objects:
\describe{
  \item{pai}{a PackedSpatRaster of monthly plant area index values, 12 layers}
  \item{hgt}{a PackedSpatRaster object of vegetation heights (m)}
  \item{x}{a PackedSpatRaster object of the ratio of vertical to horizontal projections of leaf foliage}
  \item{gsmax}{a PackedSpatRaster object maximum stomatal conductances (mol / m^2 / s)}
  \item{leafr}{a PackedSpatRaster object of leaf reflectance values}
  \item{clump}{a PackedSpatRaster of monthly values between 0 and 1 indicating the fraction of radiation passing through larger gaps in the canopy, here estimated using [clumpestimate()], 12 layers}
  \item{leafd}{a PackedSpatRaster object of leaf diameters (m)}
  \item{leaft}{a PackedSpatRaster object of leaf transmittance}
}
"vegp"

A dataset of soilparameters parameters for the point model.

An object of class groundparams

A data frame with the following columns:
\describe{
  \item{Soil.type}{description of soil type}
  \item{Smax}{Volumetric water content at saturation (m^3 / m^3)}
  \item{Smin}{Residual water content (m^3 / m^3)}
  \item{alpha}{Shape parameter of the van Genuchten model (cm^-1)}
  \item{n}{Pore size distribution parameter (dimensionless, > 1)}
  \item{Ksat}{Saturated hydraulic conductivity (cm / day)}
  \item{Vq}{Volumetric quartz fraction of soil}
  \item{Vm}{Volumetric mineral fraction of soil}
  \item{Vo}{Volumetric organic fraction of soil}
  \item{Mc}{Mass fraction of clay}
  \item{rho}{Soil bulk density (Mg / m^3)}
  \item{b}{Shape parameter for Campbell model (dimensionless, > 1)}
  \item{psi_e}{Matric potential (J / m^3)}
  \item{mult}{soil moisture model radiation coefficient}
  \item{rmu}{soil moisture model rainfall coefficient}
  \item{a}{soil moisture model deeper layer multiplier coefficient}
  \item{pwr}{soil moisture model deeper layer power coefficient}
}
}
\source{
: \url{https://onlinelibrary.wiley.com/doi/full/10.1002/ird.1751}
}
\usage{
soilparamsp
}
\description{
An object of class vegparams for the area for the area bounded by 169475, 169525, 12475, 12525
(xmin, xmax, ymin, ymax) using the Ordnance Survey GB Grid Reference system (CRS: 27700).
}
\keyword{datasets}
