% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rimplementation.R
\name{soiltemp}
\alias{soiltemp}
\title{Calculates ground surface temperature (hourly)}
\usage{
soiltemp(micro, reqhgt = 0.05, pai_a = NA, tfact = 1.5)
}
\arguments{
\item{micro}{an object of class `microin` as returned by [modelin()] (see details)}

\item{reqhgt}{height above ground at which model outputs are needed (m).}

\item{pai_a}{plant area index above `reqhgt`. Determined from total `pai` if not supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}
}
\value{
an object of class micro with the following components added:
\describe{
  \item{T0}{ground surface temperature (deg C)}
  \item{G}{ground heat flux (W/m^2)}
  \item{...}{additional terms needed for subsequent modelling}
}
}
\description{
The function `soiltemp` estimates ground surface temperature
}
\details{
if [wind()] has not been run to add additional elements to `micro`
it is automatically called.
}
\seealso{
[wind()]
}
