% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{subsetsnowmodel}
\alias{subsetsnowmodel}
\title{Subsets outputs of snow model}
\usage{
subsetsnowmodel(snowmod, subs)
}
\arguments{
\item{snowmod}{a list of model outputs from the snow model as returned by [runsnowmodel()].}

\item{subs}{a vector of index values indicating which hours to select form the snow model}
}
\description{
The function `subsetsnowmodel` subsets snow model outputs as returned
by [runsnowmodel()]
}
\examples{
# Run full snow model (takes ~90 seconds)
climdata$temp <- climdata$temp - 8 # Make it colder so there is snow
micropoint <- runpointmodel(climdata, reqhgt = 0.05, dtmcaerth, vegp, soilc) # Make it colder so there is snow
smod <- runsnowmodel(climdata, micropoint, vegp, soilc, dtmcaerth)
# Subset snow model using subset point model
smicropoint <- subsetpointmodel(micropoint, tstep = "month", what = "tmax")
smods <- subsetsnowmodel(smod, smicropoint$subs)
}
