% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{vegpfromhab}
\alias{vegpfromhab}
\title{Generate vegetation parameters form habitat type}
\usage{
vegpfromhab(
  habitats,
  hgts = NA,
  pai = NA,
  lat = NA,
  long = NA,
  tme = NA,
  clump0 = TRUE
)
}
\arguments{
\item{habitats}{a SpatRaster object of habitat types expressed as integers (see details)}

\item{hgts}{an optional SpatRaster object of vegetation heights. Estimated from habitat type if not provided.}

\item{pai}{an optional SpatRaster of plant area index values. Estimated at monthly intervals
from habitat type, with seasonal variation determined form location and date if not provided.}

\item{lat}{latitude in decimal degrees. Only needed if `pai` not provided. Estimated form `habitats` if not provided}

\item{long}{longitude in decimal degrees. Only needed if `pai` not provided. Estimated form `habitats` if not provided}

\item{tme}{POSIXlt object of dates. Only needed if `pai` not provided (see details). Taken to be current year if not provided}

\item{clump0}{optional logical, which if TRUE sets the canopy clumping factor to 0, and if false, estimates it using [clumpestimate()]}
}
\value{
an object of class vegparams - a list with the following elements:

`pai` a SpatRaster object of monthly plant area index values (see details).

`hgt` a SpatRaster object if vegetation heights (m)

`x` a SpatRaster object of ratios of vertical to horizontal projections of leaf foliage

`gsmax` a SpatRaster object of maximum stomatal conductances (mol / m^2 / s)

`leafr` a SpatRaster object of leaf reflectance values (to shortwave radiation)

`clump` a SpatRaster object of monthly values indicating the degree of canopy clumpiness, by default set to 0 (vegetation not clumped)

`leafd` a SpatRaster object of mean leaf widths (m)

`leaft` a SpatRaster object of mean leaf transmittance (m)
}
\description{
The function `vegpfromhab` generates an object of class vegparams from
a SpatRaster object of habitat types
}
\details{
This function estimates the vegetation parameters needed to run microclimf from habitat types.
Plant area index values represent the combined one sided woody and green vegetation
plant area per unit ground area. If not provided, then approximated
from habitat type, location and date. The procedure is based on calibration
against MODIS-derived estimates, accounting for regional climate. An inbuilt dataset
of regional rainfall and temperature is included with the package. Monthly
values for each unique month in `tme` are returned. Note that values are
assumed spatially constant across a given habitat type, which is unlikely
to be the case in reality. Likewise, if vegetation height values are not provided,
these are estimated form habitat type and assumed constant within that habitat type.
Habitat types should be expressed as integers as follows:
(1) for Evergreen needleleaf forest,
(2) for Evergreen broadleaf forest,
(3) for Deciduous needleleaf forest,
(4) for Deciduous broadleaf forest,
(5) for Mixed forest,
(6) for Closed shrubland,
(7) for Open shrubland,
(8) for Woody savanna,
(9) for Savanna,
(10) for Short grassland,
(11) for Tall grassland,
(12) for Permanent wetland,
(13) for Cropland,
(14) for Urban and built-up,
(15) for Cropland / Natural vegetation mosaic and
(16) for Barren or sparsely vegetated
}
\examples{
library(terra)
tme<-as.POSIXlt(c(0:8783)*3600,origin="2000-01-01 00:00", tz = "GMT")
veg<-vegpfromhab(habitats,lat=50,long=-5,tme=tme)
plot(veg$pai[[1]]), main = "Jan PAI")
plot(veg$hgt, main = "Vegetation height")
plot(veg$x, main = "Leaf angle coefficient")
plot(veg$gsmax, main = "Maximum stomatal conductance")
plot(veg$leafr, main = "Leaf reflectance")
plot(veg$clump[[1]]), main = "Canopy clumping factor")
plot(veg$leafd, main = "Leaf diameter")
plot(veg$leaft, main = "Leaf transmittance")
}
