% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{writetonc}
\alias{writetonc}
\title{Writes model output as ncdf4 file}
\usage{
writetonc(mout, fileout, dtm, weather, reqhgt, merid = 0, dst = 0)
}
\arguments{
\item{mout}{an object of class microut as returned by runmicro_hr or runmicro_dy with expand = TRUE}

\item{fileout}{output filename}

\item{dtm}{a SpatRast covering the extent of the model outputs}

\item{weather}{a data.frame of hourly weather variables (used to get time stamp)}

\item{reqhgt}{at at which model was run (m)}

\item{merid}{longitude of local time zone meridian (decimal degrees)}

\item{dst}{numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if merid = 0).}
}
\description{
The function `writetonc` writes hourly model outputs as an ncdf4 file to a specified file
}
