% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albedo}
\alias{albedo}
\title{A one m resolution dataset of white sky albedos}
\format{
A PackedSpatRaster object with 50 rows and 50 columns
}
\source{
derived using aerial imagery from \url{https://digimap.edina.ac.uk/}
}
\usage{
albedo
}
\description{
A spatial dataset of white sky albedos for Caerthillean Cove: an area bounded by 169475, 169525, 12475, 12525
(xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference system (CRS: 27700).
}
\keyword{datasets}
