% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{below_hr}
\alias{below_hr}
\title{Estimate temperature below ground (hourly)}
\usage{
below_hr(
  micro,
  reqhgt,
  xyf = NA,
  zf = NA,
  soilinit = c(NA, NA),
  tfact = 1.5,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA,
  maxhgt = NA,
  twi = NA
)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{reqhgt}{height above ground at which model outputs are needed (m). Must be negative.}

\item{xyf}{optional input for called function [wind()]}

\item{zf}{optional input for called function [wind()]}

\item{soilinit}{initial soil moisture fractions in surface and subsurface layer (see [soilmpredict()])}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{slr}{an optional RasterLayer of slope values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional RasterLayer of aspect values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from dtm if not supplied, but outer cells will be NA.}

\item{maxhgt}{an optional height (m) for which wind speed is needed. Determined
from height of tallest vegetation or as 2 m, whichever is greater, if not supplied.}

\item{twi}{optional raster of topographic wetness index values.
Calculated from `dtm` of not supplied, but outer cells will be NA.
Calculated from `dtm` of not supplied, but outer cells will be NA.}
}
\value{
`Tz` Array of air temperatures at height `reqhgt` below ground (deg C)
}
\description{
The function `below_hr` runs the below ground component of
the microclimate model at hourly time increments.
}
\seealso{
[temphumE()] for running the above ground component of the microclimate model
}
