% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{canopyrad}
\alias{canopyrad}
\title{Calculates radiation absorbed by canopy}
\usage{
canopyrad(micro, slr = NA, apr = NA, hor = NA)
}
\arguments{
\item{micro}{an object of class `micro` as returned by [modelin()]}

\item{slr}{an optional RasterLayer of slope values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional RasterLayer of aspect values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied, but outer cells will be NA.}
}
\value{
an object of class microin with the following components added:

`swabs` shortwave radiation absorbed by the canopy per unit leaf area (W/m^2)

`radm` coefficient indicating the fraction of radiation intercepted by
the canopy surface relative to horizontal
}
\description{
The function `canopyrad` calculates mean shortwave
radiation absorbed by the canopy per unit plant area
}
\details{
if [groundrad()] has not been run to add additional elements to `micro`
it is automatically called.
}
\seealso{
[groundrad()]
}
