% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{checkinputs}
\alias{checkinputs}
\title{Check format and values in model inputs}
\usage{
checkinputs(
  weather,
  rainfall,
  vegp,
  soilc,
  dtm,
  merid = 0,
  dst = 0,
  daily = FALSE
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details)}

\item{rainfall}{a vector of daily rainfall}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()] (see details)}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{dtm}{a RasterLayer onject of elevations (see details)}

\item{merid}{optionally, longitude of local time zone meridian (decimal degrees)}

\item{dst}{optionally, numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if merid = 0).}

\item{daily}{optional logical indicating whether input weather data are daily}
}
\value{
a list of checked inputs
}
\description{
The function `checkinputs` checks all the inputs into the model and returns errors
and warnings
}
\details{
The format and and units of `weather` must follow that in the example
dataset `climdata`. The array of Plant Area index values in `vegp` must
of the same x and y dims as `dtm` but can contain any number of repeated
measures up to the number of entries in `weather`. Data are interpolated to the
time increment of `weather`. Other vegetation paramaters, including vegetation
height are assumed time-invarient. The RasterLayer datasets in `soilc` must have
the same x and y dims as `dtm`. The x,y and z units of `dtm` must be all be in
metres and the coordinate reference system must be defined.
}
\examples{
# No warnings or errors given:
checks<-checkinputs(climdata, rainfall, vegp, soilc, dtmcaerth)
# Warning given
weather<-climdata
weather$relhum[1]<-101
checks<-checkinputs(weather, rainfall, vegp, soilc, dtmcaerth)
# Error given
weather<-climdata
weather$pres<-weather$pres*1000
checks<-checkinputs(weather, rainfall, vegp, soilc, dtmcaerth)
# Error given for vegp
vegp2<-vegp
vegp2$clump<-1
checks<-checkinputs(climdata, rainfall, vegp2, soilc, dtmcaerth)
# Warning given for vegp
vegp2<-vegp
vegp2$pai<-vegp$pai*10
checks<-checkinputs(climdata, rainfall, vegp2, soilc, dtmcaerth)
}
