% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climdata}
\alias{climdata}
\title{A data frame of hourly weather}
\format{
a data frame with the following elements:
\describe{
 \item{obs_time}{POSIXlt object of dates and times}
 \item{temp}{temperature (degrees C)}
 \item{relhum}{relative humidity (percentage)}
 \item{pres}{atmospheric press (kPa)}
 \item{swdown}{Total downward shortwave radiation (W / m^2)}
 \item{difrad}{Diffuse radiation (W / m^2)}
 \item{lwdown}{Total downward longwave radiation (W / m^2)}
 \item{windspeed}{Wind speed (m/s)}
 \item{winddir}{Wind direction (decimal degrees)}
 \item{precip}{Precipitation (mm)}
}
}
\usage{
climdata
}
\description{
A data frame of hourly weather in 2017 at Caerthillian Cove, Lizard, Cornwall (49.96807N, 5.215668W)
}
\keyword{datasets}
