% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{clumpestimate}
\alias{clumpestimate}
\title{Estimate canopy clumpiness}
\usage{
clumpestimate(hgt, leafd, pai, maxclump = 0.95)
}
\arguments{
\item{hgt}{a single layer SpatRaster of vegetation heights (m).}

\item{leafd}{a single layer SpatRaster of mean leaf widths (m).}

\item{pai}{a single or multi-layer SpatRaster of plant area index value(s).}

\item{maxclump}{optional parameter indicating maximum clumpiness}
}
\value{
`clump` a SpatRaster indicating the fraction of radiation passing directly
through larger gaps in the canopy. Matches the dimensions of `pai`.
}
\description{
The function `clumpestimate` estimates how clumpy a canopy is based on height and
plant area index
}
