% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{.gcanopy}
\alias{.gcanopy}
\title{Calculates turbulent molar conductivity within canopy}
\usage{
.gcanopy(l_m, a, hgt, uh, z1, z0, gmin)
}
\description{
Calculates turbulent molar conductivity within canopy
}
