% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langrangian.R
\name{.LangrangianSimT}
\alias{.LangrangianSimT}
\title{Performs Langrangian simulation (temperature)}
\usage{
.LangrangianSimT(reqhgt, micro, TH, pai_a = NA, folden = NA)
}
\description{
Performs Langrangian simulation (temperature)
}
