% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.latslonsfromr}
\alias{.latslonsfromr}
\title{lats and longs from SpatRaster object, including reprojection}
\usage{
.latslonsfromr(r)
}
\description{
lats and longs from SpatRaster object, including reprojection
}
