% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.windshelter}
\alias{.windshelter}
\title{Calculate shelter coefficient for all wind directions}
\usage{
.windshelter(wdir, dtm, whgt, s, wsa = NA)
}
\description{
Calculate shelter coefficient for all wind directions
}
