% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.windsheltera}
\alias{.windsheltera}
\title{Calculates array of wind shelter coefficients in wdct directions}
\usage{
.windsheltera(dtm, whgt, s)
}
\description{
Calculates array of wind shelter coefficients in wdct directions
}
