% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtmcaerth}
\alias{dtmcaerth}
\title{A one m resolution digital terrain dataset.}
\format{
A RasterLayer object with 50 rows and 50 columns
}
\source{
\url{http://www.tellusgb.ac.uk/}
}
\usage{
dtmcaerth
}
\description{
A dataset of elevations (m) for Caerthillean Cove: an area bounded by 169475, 169525, 12475, 12525
(xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference system (CRS: 27700).
}
\keyword{datasets}
