% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{leafrfromalb}
\alias{leafrfromalb}
\title{Derives leaf and ground reflectance from albedo}
\usage{
leafrfromalb(pai, x, alb, ltrr = 0.5, plotprogress = TRUE)
}
\arguments{
\item{pai}{a SpatRaster of plant area index values}

\item{x}{a SpatRaster of the ratio of vertical to horizontal projections of leaf foliage}

\item{alb}{a SpatRaster of white-sky albedo}

\item{ltrr}{an optional numeric value giving an approximate estimate of the ratio of leaf transmittance to leaf reflectance (e.g. value of 1 makes leaf transmittance equal to reflectance). See details}

\item{plotprogress}{optional logical indicating whether to plot results from each iteration as
model converges}
}
\value{
a list of the following
\describe{
  \item{leafr}{Leaf reflectance (range 0 - 1)}
  \item{leaft}{Leaf transmittance (range 0 - 1)}
  \item{gref}{Ground reflectance (range 0 - 1)}
}
}
\description{
Derives leaf and ground reflectance from albedo
}
\details{
the microclimate model is not unduly sensitive to `lttr` so if unknown, an approximate
value or the default can be used.
}
\examples{
pai <- rast(vegp$pai)[[9]] # Plant Area Index in Sep (month in which albedo image was flown)
x <- rast(vegp$x)
alb <- rast(albedo)
lgr <- leafrfromalb(pai, x, alb)
plot(rast(lgr$leafr))
plot(rast(lgr$gref))
}
