% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonhourly.R
\name{modelin_dy}
\alias{modelin_dy}
\title{Create object of class microindaily}
\usage{
modelin_dy(micropoint, vegp, soilc, dtm, runchecks = TRUE, xyf = 1)
}
\arguments{
\item{micropoint}{an object of class micropoint as returned by [runpointmodel()] or [subsetpointmodel()]}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()] (see details)}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{dtm}{a PackedSpatRaster or SpatRaster object of elevations (see details)}

\item{runchecks}{optional logical indicating whether to call [checkinputs()] to run
checks on format and units of input data.}

\item{xyf}{optional spatial smoothing factor applied in calculation of surface
roughness and zero-plane displacement heights (see details)}

\item{windhgt}{height above ground of wind speed measurement (m) in weather dataset (see details).}
}
\description{
The function `modelin_dy` creates an object of class microindaily
which unpacks various component inputs and reformats as required
for running the model in daily timesteps
}
\details{
see details for modelin.
}
\seealso{
[checkinputs()], [modelin()], [modelina_dy()]
}
