% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonhourly.R
\name{modelina_dy}
\alias{modelina_dy}
\title{Create object of class microindaily with weather data as an array}
\usage{
modelina_dy(
  micropointa,
  vegp,
  soilc,
  dtm,
  dtmc,
  altcorrect = 0,
  runchecks = TRUE,
  xyf = 1
)
}
\arguments{
\item{micropointa}{a list of objects of class micropoint as returned by [runpointmodela()] or [subsetpointmodela()]}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()] (see details)}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{dtm}{a SpatRaster object of elevations in metres (see details)}

\item{dtmc}{a SpatRaster object giving the resolution, spatial extent, and projection
of the weather data used when running [micropointa()]. Must give elevations in metres if
`altcorrect` > 0.}

\item{altcorrect}{a single numeric value indicating whether to apply an elevational lapse rate correction to temperatures (0 = no correction, 1 = fixed lapse rate correction, 2 = humidity-dependent variable lapse rate correction, see details)}

\item{runchecks}{optional logical indicating whether to call [checkinputs()] to run}

\item{xyf}{optional spatial smoothing factor applied in calculation of surface
roughness and zero-plane displacement heights (see [modelin()])}
}
\description{
The function `modelina_dy` creates an object of class microindaily
which unpacks various component inputs and reformats as required
for running the model in hourly timesteps. Here it is assumed that the input
weather data are as arrays - i.e. variable in space
}
\details{
see details of modelin and modelina.
}
\seealso{
[modelin()], [modelina()], [nctoclimarray()]
}
