% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{PenMont}
\alias{PenMont}
\title{Calculates Latent heat exchange}
\usage{
PenMont(
  tc,
  pk,
  ea,
  radabs,
  gHa,
  gs,
  g0,
  T0,
  es = NA,
  tdew = NA,
  surfwet = 1,
  allout = FALSE
)
}
\arguments{
\item{tc}{air temperature (deg c)}

\item{pk}{atmospheric pressure (kPa)}

\item{ea}{vapour pressure of air (kPa)}

\item{radabs}{absorbed long and shortwave radiation (W/m2)}

\item{gHa}{boundary layer conductances for heat (mol/m^2/s)}

\item{gs}{stomatal conductance (mol/m^2/s)}

\item{g0}{conductance to ground}

\item{T0}{ground surface temperature (deg C)}

\item{es}{saturated vapour pressure (kPa) at temperature tc (calculated if NA)}

\item{tdew}{dewpoint temperature (deg C) (calculated if NA, see details)}

\item{surfwet}{proportion of leaf or canopy acting as wet surface}

\item{allout}{optional logical indicating whether to return Latent, Sensible and
Ground heat flux in addition to canopy temperature}
}
\value{
a list with the following components:

`tcan` the leaf or effective canopy temperature (deg C)

`H` the sensible heat flux (w/m^2)

Optionally `L` The latent heat flux (W/m^2)

Optionally `G` The ground heat flux (W/m^2)
}
\description{
The function `PenMont` applies a varient of the Penman-Monteith
equation, with G calculated from ground temperatures, to calculate latent
heat exchange
}
\details{
`Latent` can be used to calaculate either the latent heat of
evapotranspiration from individual leaves or bulk surface latent heat of
evapotranspiration from the canopy as a whole depending on how `gHa` and `gs`
are defined. If for the canopy as a whole, heat storage `G` should also be
considered and `L` is the Latent heat per unit leaf area and should thus be
multiplied by the total leaf area per unit ground area to derive a total. If
for a leaf in isolation, set g0 to zero and T0 to an arbitrary value. The
paramater `surfwet` determines how much of the canopy should be treated as
wet surface. However, except when extremely droughted, the matric potential of
leaves is such that `surfwet` ~ 1.
}
