% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimfunctions.R
\name{runbioclim}
\alias{runbioclim}
\title{Runs microclimate model to produce bioclim variables}
\usage{
runbioclim(
  weather,
  precip,
  tme = NA,
  reqhgt = 0.05,
  micropoint = NA,
  vegp,
  soilc,
  dtm,
  dtmc = NA,
  temp = "air",
  hourly = TRUE,
  altcorrect = 1,
  windhgt = 2,
  soilm = NA,
  dTmx = 25,
  maxiter = 100,
  runchecks = TRUE,
  xyf = 1,
  pai_a = NA,
  tfact = 1.5,
  surfwet = NA,
  slr = NA,
  apr = NA,
  hor = NA,
  twi = NA,
  wsa = NA,
  lat = NA,
  long = NA
)
}
\arguments{
\item{weather}{a data.frame or list of arrays of weather variables (as for [runpointmodel() or [runpointmodela()])}

\item{precip}{a vector or array of daily precipitation variables ((as for [runpointmodel() or [runpointmodela()]))}

\item{tme}{If weather is not a data.frame, an object of class POSIXlt giving the dates and times for each weather variable stored in the array. Set to NA if weather is a data,frame}

\item{reqhgt}{height for which temperatures are needed}

\item{micropoint}{optional subset output of point microclimate model for days with monthly monthly median, maximum and minimum temperatures. Calculated if not supplied.}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()]}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{dtm}{a SpatRaster object of elevations in metres}

\item{dtmc}{If weather is s not a data.frame, a SpatRaster object giving the resolution,
spatial extent, and projection of the weather data. Must give elevations in meters if
`altcorrect` > 0.}

\item{temp}{one of "air" or "leaf" indicating whether bioclim variables are constructed using air or leaf
temperatures. If `temp = leaf`, for grid cells where `reqhgt` is above vegetation, vertically averaged canopy
temperature is used.}

\item{hourly}{optional logical indicating whether to run the model in hourly or daily mode.}

\item{altcorrect}{a single numeric value indicating whether to apply an elevational lapse rate correction to temperatures (0 = no correction, 1 = fixed lapse rate correction, 2 = humidity-dependent variable lapse rate correction)}

\item{windhgt}{height above ground of wind speed data in weather}

\item{soilm}{optional vector of soil moisture values in upper 10 cm of the soil (calculated if not supplied)}

\item{dTmx}{optional maximum amount by which canopy or ground surface temperatures can exceed air temperatures.
Included to ensure model convergence.}

\item{maxiter}{optional integer indicating the maximum number of iterations (see details).}

\item{runchecks}{optional logical indicating whether to call [checkinputs()] to run}

\item{xyf}{optional spatial smoothing factor applied in calculation of surface
roughness and zero-plane displacement heights (see [modelin()])}

\item{pai_a}{an optional array of plant area index values above `reqhgt`.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation in topographic wetness (see [soilmdistribute()]).}

\item{surfwet}{an optional single numeric value of array of values specifying the proportion
of the canopy surface that should be treated as wet surface (modelled if not supplied)}

\item{slr}{slr an optional SpatRaster object of slope values (Radians). Calculated from
`dtm` if not supplied, but outer cells will be NA.}

\item{apr}{an optional SpatRaster object of aspect values (Radians). Calculated from
`dtm` if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from `dtm` if not supplied, but outer cells will be NA.}

\item{twi}{optional SpatRaster object of topographic wetness index values.
Calculated from `dtm` if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from `dtm` if not supplied, but outer cells will be NA.}

\item{lat}{optional central latitude of study area (removes tile effects when running in tiles)}

\item{long}{optional central longitude of study area (removes tile effects when running in tiles)}
}
\value{
a multilayer SpatRast of the following:
\describe{
 \item{BIO1}{Mean of monthy median temperatures (degrees C)}
 \item{BIO2}{Mean diurnal temperature range (degrees C)}
 \item{BIO3}{Isothermality (BIO2/BIO7) (×100)}
 \item{BIO4}{Temperature Seasonality (standard deviation of monthly median temperatures × 100)}
 \item{BIO5}{Maximum temperature (degrees C)}
 \item{BIO6}{Minimum temperature (degrees C)}
 \item{BIO7}{Temperature Annual Range (BIO5-BIO6) (degrees C)}
 \item{BIO8}{Mean of monthly median temperatures in wettest three months (degrees C)}
 \item{BIO9}{Mean of monthly median temperatures in driest three months (degrees C)}
 \item{BIO10}{Mean of monthly median temperatures in warmest three months (degrees C)}
 \item{BIO11}{Mean of monthly median temperatures in coldest three months (degrees C)}
 \item{BIO12}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of each month (m^3 / m^3)}
 \item{BIO13}{Maximum of monthly soil moistures on the hottest, coldest and median temperature day of each month (m^3 / m^3)}
 \item{BIO14}{Minimum of monthly soil moistures on the hottest, coldest and median temperature day of each month (m^3 / m^3)}
 \item{BIO15}{Soil moisture seasonality (Coefficient of Variation) on day in each month with median temperature}
 \item{BIO16}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of the wettest three months (m^3 / m^3)}
 \item{BIO17}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of the driest three months (m^3 / m^3)}
 \item{BIO18}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of the warmest three months (m^3 / m^3)}
 \item{BIO19}{Mean of monthly soil moistures on the hottest, coldest and median temperature day of the coldest three months (m^3 / m^3)}
}
}
\description{
The function `runbioclim` runs the microclimate model and produces microclimate
equivalents of the 19 Worldclim bioclimate variables
}
\details{
To enhance computational efficiency the microclimate model is run for selected days only. Thus, to compute
mean annual temperature, the mean ambient temperature of each day in the input weather data is calculated,
the day with median temperatures in each month selected and the mean across months calculated. This is
not, strictly speaking, the same as the mean temperature, but differences are likely to be minor, and for each year of
data supplied, there is an approximately 30-fold gain in computational efficiency by calculating
BIO1 in this way. Similarly, to calculate maximum temperature (BIO5), the day of the year with the hottest ambient
temperature is selected, and microclimate temperatures calculated on this day only. This ignores
the possibility that on a slightly cooler, but sunnier day, microclimate temperatures may be hotter
hotter at certain locations. If `hourly = TRUE` all hours within a given day are selected and calculations
performed on hourly data. If If `hourly = FALSE` only the hours corresponding to times when hourly temperatures
are at their daily maximum and minimum and selected. This results in a c. 10-fold increase in computational
efficiency, but cannot pick out areas where terrain results in near-ground temperatures reaching
a maximum later in the afternoon than the peak in ambient temperature. If weather data for more than one year are supplied,
only one set of median, maximum and minimum monthly temperature data are selected representing an average
across years. Resultant, there is little computational penalty if providing data for multiple years
in comparison to one year of data.
}
\examples{
# Takes ~20 seconds to run
bioclim<-runbioclim(climdata, rainfall, tme = NA, reqhgt = 0.05, micropoint = NA,
                    vegp, soilc, dtmcaerth)
plot(bioclim[[1]]) # BIO1
}
