% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cppwrappers.R
\name{runmicro_big}
\alias{runmicro_big}
\title{runmicro on big areas}
\usage{
runmicro_big(
  micropoint,
  reqhgt,
  pathout = getwd(),
  vegp,
  soilc,
  dtm,
  dtmc = NA,
  altcorrect = 0,
  tilesize = NA,
  toverlap = 0,
  writeasnc = FALSE,
  runchecks = TRUE,
  pai_a = NA,
  tfact = 1.5,
  out = rep(TRUE, 10)
)
}
\arguments{
\item{micropoint}{an object of class micropoint or a list of objects of class
micropoint as returned by [runpointmodel()], [runpointmodela()], [subsetpointmodel()]
or [subsetpointmodela()]}

\item{reqhgt}{height above (postive) or below (negative) ground for which microclimate variables are required (m)}

\item{pathout}{a file directory to which to save data. Data saved to a subdirectory
called `microut` in this directory. Default is working directory.}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()] (see details)}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{dtm}{a SpatRaster object of elevations in metres (see details)}

\item{dtmc}{a SpatRaster object giving the resolution, spatial extent, and projection
of the climate data used when running [micropointa()]. Ignored if climate data
used for running the point model are provided as a data.frame. Must
give elevations in metres if `altcorrect` > 0 or if setting runchecks to TRUE.}

\item{altcorrect}{a single numeric value indicating whether to apply an elevational lapse rate correction to temperatures (0 = no correction, 1 = fixed lapse rate correction, 2 = humidity-dependent variable lapse rate correction, see details)}

\item{tilesize}{optional integer of the number of pixels in x and y of each tile (returned tiles are square).
Calculated automatically based on data size of not provided.}

\item{toverlap}{optional integer specifying the number of pixels of overlap between adjacent tiles. Default 0,
Set to > 0 and use [mosaicblend()] if output appears to have tiling effects.}

\item{writeasnc}{optional logical indicating whether to write output data as netCDF4 files (default TRUE). Can only be used
if all entries of `out` are true. Alternatively data are saved as RDS files.}

\item{runchecks}{optional logical indicating whether to call [checkinputs()] to run}

\item{pai_a}{an array of plant area index values above `reqhgt`. Determined from total `pai` if not supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{out}{optional vector of logicals indicating which variables to
return ordered as for the listed outputs when `reqhgt > 0'` (e.g. `out[1] = TRUE` indicates that
`Tz` is returned, `out[2]` that `tleaf` is returned etc). By default all variables
are returned.}

\item{silent}{optional logical indicating whether to report on progress (default FALSE - progress reported).}
}
\description{
The function `runmicro` runs the grid version of the microclimate model
}
\details{
The function `runmicro_big` tiles larger studies and saves outputs for each tile

if `writeasnc = TRUE` and all entries in `out` are `TRUE`, a ncdf4
files for each tile are written out, unless the corresponding tile comprises
all NAs. To save memory, data are written out as integers: temperatures and wind
speeds are multiplied by 100 prior to doing so. If `writeasnc = FALSE` or
not all variables are required, then an RDS file is written out. a dtm (as a
rwapped SpatRaster) of the corresponding tile is attached to the model outputs prior to doing so to
aid with georeferencing.
see [runmicro()]
}
