% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{runmicro_big}
\alias{runmicro_big}
\title{runmicro on big areas}
\usage{
runmicro_big(
  weather,
  rainfall,
  reqhgt,
  vegp,
  soilc,
  dtm,
  pathout,
  hourly = FALSE,
  pai_a = NA,
  folden = NA,
  xyf = 1,
  zf = NA,
  soilinit = c(NA, NA),
  tfact = 1.5,
  surfwet = 1,
  merid = 0,
  dst = 0,
  runchecks = TRUE
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details)}

\item{rainfall}{a vector of daily rainfall}

\item{reqhgt}{height above ground at which model outputs are needed (m).}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()] (see details)}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{dtm}{a RasterLayer onject of elevations (see details)}

\item{pathout}{a file directory to which to save data}

\item{hourly}{optional logical indicating whether to expand model to hourly and write
outputs as ncdf4 files, or keep outputs at daily and save raw outputs}

\item{pai_a}{an optional array of plant area index values above `reqhgt` (see details)}

\item{xyf}{optional input for called function [wind()]}

\item{zf}{optional input for called function [wind()]}

\item{soilinit}{initial soil moisture fractions in surface and subsurface layer (see [soilmpredict()])}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{surfwet}{an optional single numeric value of array of values specifying the proportion
of the canopy surface that should be treated as wet surface (see details)}

\item{merid}{optionally, longitude of local time zone meridian (decimal degrees)}

\item{dst}{optionally, numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if merid = 0).}
}
\value{
if `hourly` = TRUE, ncdf4 files of hourly values for each 100 x 100 grid cell tiles of
the study area, numbered by row and column, and saved in a folder `microut` in the
directory specified by path.

if `hourly` = FALSE, daily max and min values saved in files with a .R
extension for each 100 x 100 grid cell tiles of the study area, numbered by row
and column, and saved in a folder `microut` in the directory specified by path. This
saves disk space, is singnificantly faster and files can later be expanded to
hourly and written out as ncdf4 files using [expandtonc()]
}
\description{
The function `runmicro_big` tiles larger studies and saves outputs for each tile
}
\details{
The parmater `pai_a` is used to calaculate the radiation incercepted by leaves at `reqhgt` if
below canopy. If not supplied it is calculated from total plant area index by
assuming leaf density within the canopy is uniformly vertically distributed. If suuplied
it must have the same dimensions as micro$pai. I.e. with the same x and y dims as the
the supplied dtm and values for each hour as the z dimension. The paramater `surfwet`
determines how much of the canopy should be treated as wet surface when calaculating
latent heat fluxes. However, except when extremely droughted, the matric potential of leaves
is such that `surfwet` ~ 1.
}
