% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{runmicro_biga}
\alias{runmicro_biga}
\title{runmicro on big areas with climate data provided as arrays}
\usage{
runmicro_biga(
  climarray,
  precarray,
  tme,
  r,
  altcorrect,
  reqhgt,
  vegp,
  soilc,
  dtm,
  pathout,
  hourly = FALSE,
  windhgt = 2,
  pai_a = NA,
  folden = NA,
  xyf = 1,
  zf = NA,
  soilinit = c(NA, NA),
  tfact = 1.5,
  surfwet = 0.75,
  merid = 0,
  dst = 0,
  runchecks = TRUE,
  soilmcoefs = NA,
  soiltcoefs = NA,
  backweight = 2,
  maxiter = 100,
  gmn = 0.3
)
}
\arguments{
\item{climarray}{a list of arrays of weather variables (see details). See also [nctoarray()]}

\item{tme}{an object of class POSIXlt giving the dates and times for each weather variable stroed in the array}

\item{r}{a SpatRaster object giving with the resolution, spatial extent, and projection of the weather data (see details)}

\item{altcorrect}{a single numeric value indicating whether to apply an elevational lapse rate correction to temperatures (0 = no correction, 1 = fixed lapse rate correction, 2 = humidity-dependent variable lapse rate correction, see details)}

\item{reqhgt}{height above ground at which model outputs are needed (m).}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()] (see details)}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{dtm}{a SpatRast or PackedSpatRast object of elevations (see details)}

\item{pathout}{a file directory to which to save data}

\item{hourly}{optional logical indicating whether to expand model to hourly and write
outputs as ncdf4 files, or keep outputs at daily and save raw outputs}

\item{windhgt}{height above ground of wind speed measurement (m) in climate dataset}

\item{pai_a}{an optional array of plant area index values above `reqhgt` (see details)}

\item{xyf}{optional input for called function [wind()]}

\item{zf}{optional input for called function [wind()]}

\item{soilinit}{initial soil moisture fractions in surface and subsurface layer (see [soilmpredict()])}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{surfwet}{an optional single numeric value of array of values specifying the proportion
of the canopy surface that should be treated as wet surface (see details)}

\item{merid}{optionally, longitude of local time zone meridian (decimal degrees)}

\item{dst}{optionally, numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if merid = 0).}

\item{precip}{an array of daily precipitation (see details)}

\item{seealso}{[runmicro_biga()] for running the microclimate model over large areas
as tiles with climate input data provided as a data.frame.}
}
\value{
if `hourly` = TRUE, ncdf4 files of hourly values for each 100 x 100 grid cell tiles of
the study area, numbered by row and column, and saved in a folder `microut` in the
directory specified by path. Data are stored as intigers and multiplied by 100

if `hourly` = FALSE, daily max and min values saved in files (using saveRDS) with a .R
extension for each 100 x 100 grid cell tiles of the study area, numbered by row
and column, and saved in a folder `microut` in the directory specified by path. This
saves disk space, is significantly faster and files can later be expanded to
hourly and written out as ncdf4 files using [expandtonc()]. Data are stored as integers and
multiplied by 100. A cropped dtm of the tile is also saved as a PackedSpatRast object.
}
\description{
The function `runmicro_biga` tiles larger studies and saves outputs for each tile
}
\details{
The parameter `pai_a` is used to calculate the radiation intercepted by leaves at `reqhgt` if
below canopy. If not supplied it is calculated from total plant area index by
assuming leaf density within the canopy is uniformly vertically distributed. If suuplied
it must have the same dimensions as micro$pai. I.e. with the same x and y dims as the
the supplied dtm and values for each hour as the z dimension. The parameter `surfwet`
determines how much of the canopy should be treated as wet surface when calculating
latent heat fluxes. However, except when extremely droughted, the matric potential of leaves
is such that `surfwet` ~ 1.
}
