% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{runmicro_dy}
\alias{runmicro_dy}
\title{Run microclimate model (daily)}
\usage{
runmicro_dy(
  microd,
  reqhgt,
  expand = TRUE,
  pai_a = NA,
  xyf = 1,
  zf = NA,
  soilinit = c(NA, NA),
  tfact = 1.5,
  surfwet = 1,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA,
  maxhgt = NA,
  twi = NA
)
}
\arguments{
\item{microd}{object of class microindaily as returned by [modelin_dy()]}

\item{reqhgt}{height above ground at which model outputs are needed (m).}

\item{expand}{optional logical indicating whether to expand daily values to hourly (see details).}

\item{pai_a}{an optional array of plant area index values above `reqhgt` (see details)}

\item{xyf}{optional input for called function [wind()]}

\item{zf}{optional input for called function [wind()]}

\item{soilinit}{initial soil moisture fractions in surface and subsurface layer (see [soilmpredict()])}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{surfwet}{an optional single numeric value of array of values specifying the proportion
of the canopy surface that should be treated as wet surface (see details)}

\item{slr}{an optional RasterLayer of slope values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional RasterLayer of aspect values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from dtm if not supplied, but outer cells will be NA.}

\item{maxhgt}{an optional height (m) for which wind speed is needed. Determined
from height of tallest vegetation or as 2 m, whichever is greater, if not supplied.}

\item{twi}{optional raster of topographic wetness index values.
Calculated from `dtm` of not supplied, but outer cells will be NA.}
}
\value{
if expand = TRUE, an object of class microout with the following components:

`Tz` Array of air temperatures at height `reqhgt` (deg C). Identical to `T0`
if `reqhgt` = 0.

`tleaf` Array of leaf temperatures at height `reqhgt` (deg C).
NA if `reqhgt` greater than canopy height or `reqhgt` <= 0.

`T0` Array of ground surface temperatures (deg C)

`relhum` Array of relative humidities at height `reqhgt` (percentage).
NA if `reqhgt` <= 0.

`windspeed` Array of wind speeds at height `reqhgt` (m/s).
NA if `reqhgt` <= 0.

`raddir` Array of direct shortwave radiation received on horizontal
surface (W/m^2). NA if `reqhgt` <= 0.

`raddif` Array of direct shortwave radiation received on horizontal
surface (W/m^2). NA if `reqhgt` <= 0.

`radlw` Array of downward longwave radiation received on horizontal
surface (W/m^2). NA if `reqhgt` <= 0.

if expand = FALSE, an object of class microutdaily, list with the following components:

mout_mn an object of class microut for minimum daily temperatures

mout_mx an object of class microut for maximum daily temperatures
}
\description{
The function `runmicro_dy` runs the microclimate model in daily
time increments, with the option to expand to hourly
}
\details{
If expand = TRUE, daily minima and maxima are expanded to hourly using values in the hourly
weather dataset. The parmater `pai_a` is used to calaculate the radiation incercepted by leaves at `reqhgt` if
below canopy. If not supplied it is calaculated from total plant area index by
assuming leaf density within the canopy is uniformly vertically distributed. If suuplied
it must have the same dimensions as micro$pai. I.e. with the same x and y dims as the
the supplied dtm and values for each hour as the z dimension. The paramater `surfwet`
determines how much of the canopy should be treated as wet surface when calaculating
latent heat fluxes. However, except when extremely droughted, the matric potential of leaves
is such that `surfwet` ~ 1.
}
\examples{
library(raster)
library(sp)
library(zoo)
library(abind)
# Create model input with inbuilt datasets
microd<-modelin_dy(climdata,rainfall,vegp,soilc,dtmcaerth)
# Run model 5 cm above ground  performing hourly expansion (takes ~50 seconds to run)
mout<-runmicro_dy(microd, 0.05)
# Plot air temperatures on hottest hour
plot(raster(mout$Tz[,,4094]))
# Plot mean air temperatures
mairt<-apply(mout$Tz,c(1,2),mean)
plot(raster(mairt))
# Plot ground temperatures on hottest and coldest hour
plot(raster(mout$T0[,,991])) # coldest hour
plot(raster(mout$T0[,,4094])) # hottest hour
# Run model 20 cm below ground
mout<-runmicro_dy(microd, -0.2)
# Extract and plot mean soil temperatures
msoilt<-apply(mout$Tz,c(1,2),mean)
plot(raster(msoilt))
}
\seealso{
[runmicro_hr()] for running microclimate model in hourly time-steps
}
