% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{runmicro_hr}
\alias{runmicro_hr}
\title{Run microclimate model (hourly)}
\usage{
runmicro_hr(
  micro,
  reqhgt,
  pai_a = NA,
  folden = NA,
  xyf = 1,
  zf = NA,
  soilinit = c(NA, NA),
  tfact = 1.5,
  surfwet = 1,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA,
  maxhgt = NA,
  twi = NA
)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{reqhgt}{height above ground at which model outputs are needed (m).}

\item{pai_a}{an optional array of plant area index values above `reqhgt` (see details)}

\item{folden}{an optional array of foliage densities at the height `reqhgt` (m^3/m^3)}

\item{xyf}{optional input for called function [wind()]}

\item{zf}{optional input for called function [wind()]}

\item{soilinit}{initial soil moisture fractions in surface and subsurface layer (see [soilmpredict()])}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{surfwet}{an optional single numeric value of array of values specifying the proportion
of the canopy surface that should be treated as wet surface (see details)}

\item{slr}{an optional SpatRaster object of slope values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional SpatRaster object of aspect values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from dtm if not supplied, but outer cells will be NA.}

\item{maxhgt}{an optional height (m) for which wind speed is needed. Determined
from height of tallest vegetation or as 2 m, whichever is greater, if not supplied.}

\item{twi}{optional SpatRaster object of topographic wetness index values.
Calculated from `dtm` of not supplied, but outer cells will be NA.}
}
\value{
an object of class microout with the following components:

`Tz` Array of air temperatures at height `reqhgt` (deg C). Identical to `T0`
if `reqhgt` = 0.

`tleaf` Array of leaf temperatures at height `reqhgt` (deg C).
NA if `reqhgt` greater than canopy height or `reqhgt` <= 0.

`T0` Array of ground surface temperatures (deg C)

`relhum` Array of relative humidities at height `reqhgt` (percentage).
NA if `reqhgt` <= 0.

`windspeed` Array of wind speeds at height `reqhgt` (m/s).
NA if `reqhgt` <= 0.

`raddir` Array of direct shortwave radiation received on horizontal
surface (W/m^2). NA if `reqhgt` <= 0.

`raddif` Array of direct shortwave radiation received on horizontal
surface (W/m^2). NA if `reqhgt` <= 0.

`radlw` Array of downward longwave radiation received on horizontal
surface (W/m^2). NA if `reqhgt` <= 0.
}
\description{
The function `runmicro_hr` runs the microclimate model in hourly
time increments
}
\details{
`pai_a` is used to calaculate the radiation incercepted by leaves at `reqhgt` if
below canopy. If not supplied it is calaculated from total plant area index by
assuming leaf density within the canopy is uniformly vertically distributed. If suuplied
it must have the same dimensions as micro$pai. I.e. with the same x and y dims as the
the supplied dtm and values for each hour as the z dimension. The paramater `surfwet`
determines how much of the canopy should be treated as wet surface when calaculating
latent heat fluxes. However, except when extremely droughted, the matric potential of leaves
is such that `surfwet` ~ 1.
}
\examples{
library(terra)
library(zoo)
library(abind)
# Create model input with inbuilt datasets
micro<-modelin(climdata,rainfall,vegp,soilc,dtmcaerth)
# Run model 5 cm above ground (takes ~ 3 mins to run on 50 x 50 x 8760 values)
mout<-runmicro_hr(micro, 0.05)
# Plot air temperatures on hottest hour
plot(rast(mout$Tz[,,4094]))
# Plot mean air temperatures
mairt<-apply(mout$Tz,c(1,2),mean)
plot(rast(mairt))
# Plot ground temperatures on hottest and coldest hour
plot(rast(mout$T0[,,991])) # coldest hour
plot(rast(mout$T0[,,4094])) # hottest hour
# Run model 20 cm below ground
mout<-runmicro_hr(micro, -0.2)
# Extract and plot mean soil temperatures
msoilt<-apply(mout$Tz,c(1,2),mean)
plot(rast(msoilt))
}
\seealso{
[runmicro_dy()] for faster running microclimate model in daily time-steps,
including the option to expand daily outputs to hourly using the input diurnal
temperature cycle and [runmicro_big()] for running the microclimate model over large areas
as tiles.
}
