% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{runmicro_hr}
\alias{runmicro_hr}
\title{Run microclimate model (hourly)}
\usage{
runmicro_hr(
  micro,
  reqhgt,
  pai_a = NA,
  tfact = 1.5,
  surfwet = NA,
  slr = NA,
  apr = NA,
  hor = NA,
  twi = NA,
  wsa = NA
)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{reqhgt}{height above (or below) ground at which model outputs are needed (m). Negative values indicate below ground surface.}

\item{pai_a}{an optional array of plant area index values above `reqhgt` (see details)}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{surfwet}{an optional single numeric value of array of values specifying the proportion
of the canopy surface that should be treated as wet surface (see details)}

\item{slr}{an optional SpatRaster object of slope values (Radians). Calculated from
the dtm in micro if not supplied, but outer cells will be NA.}

\item{apr}{an optional SpatRaster object of aspect values (Radians). Calculated from
the dtm in micro if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from the dtm in micro if not supplied, but outer cells will be NA.}

\item{twi}{optional SpatRaster object of topographic wetness index values.
Calculated from the dtm in micro of not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from the dtm in micro if not supplied, but outer cells will be NA.}
}
\value{
`Tz` Array of air temperatures at height `reqhgt` (deg C). Identical to `T0`
if `reqhgt` = 0.

`tleaf` Array of leaf temperatures at height `reqhgt` (deg C).
NA if `reqhgt` greater than canopy height or `reqhgt` <= 0.

`T0` Array of ground surface temperatures (deg C)

`relhum` Array of relative humidities at height `reqhgt` (percentage).
NA if `reqhgt` <= 0.

`windspeed` Array of wind speeds at height `reqhgt` (m/s).
NA if `reqhgt` <= 0.

`Rdirdown` Array of downward direct shortwave radiation incident on
horizontal surface (W/m^2)

`Rdifdown` Array of downward diffuse shortwave radiation incident on
horizontal surface (W/m^2)

`Rlwdown` Array of downward longwave radiation incident on horizontal
surface (W/m^2)

`Rswup` Array of upward shortwave radiation (assumed diffuse) incident
on underside of horizontal surface (W/m^2)

`Rlwup` Array of upward longwave radiation incident on underside of
horizontal surface (W/m^2)
}
\description{
The function `runmicro_hr` runs the microclimate model in hourly
time increments
}
\details{
`pai_a` is used to calculate the radiation intercepted by leaves at `reqhgt` if
below canopy. If not supplied it is calculated from total plant area index by
assuming a realistic shape to the vertical profile foliage within the canopy. If supplied,
`pai_a` must have the same dimensions as micro$pai. I.e. with the same x and y
dims as the the supplied dtm and values for each hour as the z dimension. The parameter `surfwet`
determines how much of the canopy should be treated as wet surface when calculating
latent heat fluxes. However, except when extremely droughted, the matric potential of leaves
is such that `surfwet` ~ 1. If set to NA, surfess wetness is modelled.
}
\examples{
library(terra)
library(zoo)
library(abind)
# ** First run point model (NB not run as uses inbuilt dataset)
# micropoint<-runpointmodel(climdata, rainfall, reqhgt = 0.05, vegp, soilc)
# ** Subset inbuilt point model to get hottest days in each month
micropoint_mx<-subsetpointmodel(micropoint, tstep = "month", what = "tmax")
micropoint_mn<-subsetpointmodel(micropoint, tstep = "month", what = "tmin")
# ** Create model inputs with inbuilt datasets
micro_mx<-modelin(micropoint_mx,vegp,soilc,dtmcaerth)
micro_mn<-modelin(micropoint_mn,vegp,soilc,dtmcaerth)
# Run model 5 cm above ground
mout_mx<-runmicro_hr(micro_mx, 0.05)
mout_mn<-runmicro_hr(micro_mn, 0.05)
# Plot air temperatures on hottest hour
plot(rast(mout_mx$Tz[,,134]))
# Plot mean of monthly max and min
mairt<-apply((mout_mn$Tz + mout_mx$Tz) / 2, c(1,2),mean)
plot(rast(mairt))
}
\seealso{
[runmicro_dy()] for faster running microclimate model in daily time-steps,
including the option to expand daily outputs to hourly using the input diurnal
temperature cycle and [runmicro_big()] for running the microclimate model over large areas
as tiles.
}
