% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{runpointmodel}
\alias{runpointmodel}
\title{Runs point microclimate model}
\usage{
runpointmodel(
  weather,
  precip,
  vegp,
  soilc,
  windhgt = 2,
  reqhgt = 0.1,
  soilm = NA,
  dTmx = 25,
  maxiter = 100
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details)}

\item{precip}{a vector of daily precipitation}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()] (see details)}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{windhgt}{height above ground of wind speed data in weather}

\item{reqhgt}{height for which temperatures are needed (used only when reqhgt < 0 to calculate tmeperature below ground)}

\item{soilm}{optional vector of soil moisture values in upper 10 cm of the soil (calculated if not supplied)}

\item{dTmx}{optional maximum amount by which canopy or ground surface temperatures can exceed air temperatures.
Included to ensure model convergence}

\item{maxiter}{optional integer indicating the maximum number of iterations (see details)}
}
\value{
a list of the following:
(1) weather - a data.frame of weather variables, but with temperature and
wind speed height-adjusted to be above canopy if necessary (see details)
(2) precip - a vector of daily precipitation (same as input)
(3) microp - an object of class pointmicro as returned bgy [microiter::RunBigLeaf()].
(4) soilm - a vector of hourly soil moisture fractions in the upper 10 cm of the soil
(5) Tbz - a vector of below ground temperatures. NA if reqhgt > 0
(5) vegp - a list of vegetation paremeters used by the point model
(6) groundp - a list of ground paremeters used by the point model
(7) soiltype - soil type assumed when running the point model
(7) lat - latitude (decimal degrees) of study area centre. Determined from input raster
(8) long - longitude (decimal degrees) of study area centre. Determined from input raster
(10) tstep - here set to `hour` (see [subsetpointmodel()])
}
\description{
The function `runpointmodel` runs the point microclimate model
}
\details{
The format and and units of `weather` must follow that in the example
dataset `climdata`. As not all wind measurements are at reference height, the height of the wind speed measurement
must be specified if not 2 m. To enable calculation of below-canopy wind and temperature profiles
in tall canopy, the wind speed and temperature data are  adjusted to give values for a height at
the maximum vegetation height if the tallest vegetation exceeds two metres.
For doing so a stand vegetation surface typical of that in which a weather station would be located is assumed.
The parameter `maxiter` sets the maximum number of times the model is iterated to achieve
convergence. Increasing this value improves accuracy at the expense of computation time.
The array of Plant Area index values and clumping factors in `vegp` must
have the same x and y dims as `dtm` but can contain any number of repeated
measures up to the number of entries in `weather`. Data are interpolated to the
time increment of `weather`. Other vegetation paramaters are assumed time-invarient.
}
\examples{
# example code
# Run model:
micropoint<-runpointmodel(climdata,rainfall,vegp,soilc)
# Plot canopy heat exchange surface temperature
microp<-micropoint$microp
plot(microp$Tc,type="l") # temperature of canopy surface
}
