% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cppwrappers.R
\name{runpointmodel}
\alias{runpointmodel}
\title{Runs point microclimate model}
\usage{
runpointmodel(
  weather,
  reqhgt = 0.05,
  dtm,
  vegp,
  soilc,
  runchecks = TRUE,
  zref = 2,
  windhgt = zref,
  soilm = NA,
  matemp = NA,
  dTmx = 25,
  maxiter = 20,
  yearG = TRUE,
  lat = NA,
  long = NA,
  vegp_p = NA,
  groundp_p = NA,
  soiltype = NA,
  mxhgt = NA
)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details)}

\item{reqhgt}{height for which temperatures are needed (used only when reqhgt < 0 to calculate tmeperature below ground)}

\item{dtm}{a SpatRaster of elevations for the study area}

\item{vegp}{an object of class vegparams as returned by [vegpfromhab()] (see details)}

\item{soilc}{an object of class soilcharac as returned by [soilcfromtype()]}

\item{zref}{height above ground (m) of temperature measurements in weather}

\item{windhgt}{height above ground (m) of wind speed data in weather}

\item{soilm}{optional vector of soil moisture values in upper 10 cm of the soil (calculated if not supplied)}

\item{matemp}{optionally mean annual temperature. Only used for refining below-ground temperature estimates and
calculate as the mean of weather data temperatures if not provided.}

\item{dTmx}{optional maximum amount by which canopy or ground surface temperatures can exceed air temperatures.
Included to ensure model convergence:}

\item{maxiter}{optional integer indicating the maximum number of iterations (see details)}

\item{yearG}{optional logical indicating whether or not to include annual ground heat flux cycle
Reduces tile effects:}

\item{lat}{optional central latitude of study area (removes tile effects when running in tiles)}

\item{long}{optional central longitude of study area (removes tile effects when running in tiles)
Used by runpointmodela:}

\item{vegp_p}{optional vector of point model vegetation parameters}

\item{groundp_p}{= optional vector of point model ground parameters}

\item{soiltype}{= optional modal soil type}

\item{mxhgt}{= optional maximum height of vegetation}
}
\value{
a list of the following:
(1) weather - a data.frame of weather variables, but with temperature and
wind speed height-adjusted to be above canopy if necessary (see details)
(2) dfo - a data.frame of microclimate point model outputs required for running grid model
(3) if reqhgt < 0, a vector iof point model temperatures below ground
(4) lat - the latitude of the centroid of the study area for which the point model was run (decimal degrees)
(5) long - the longitude of the centroid of the study area for which the point model was run (decimal degrees)
(6) zref - the height to which weather data have been height adjusted (see details)
(7) subs & tmeorig used subsequently by grid model to handle time-variant vegetation inputs
}
\description{
The function `runpointmodel` runs the point microclimate model
}
\details{
The format and and units of `weather` must follow that in the example
dataset `climdata`. As not all wind measurements are at reference height, the height of the wind speed measurement
must be specified if not 2 m. To enable calculation of below-canopy wind and temperature profiles
in tall canopy, the wind speed and temperature data are  adjusted to give values for a height at
the maximum vegetation height if the tallest vegetation exceeds two metres.
For doing so a stand vegetation surface typical of that in which a weather station would be located is assumed.
The parameter `maxiter` sets the maximum number of times the model is iterated to achieve
convergence. Increasing this value improves accuracy at the expense of computation time.
The array of Plant Area index values and clumping factors in `vegp` must
have the same x and y dims as `dtm` but can contain any number of repeated
measures up to the number of entries in `weather`. Data are interpolated to the
time increment of `weather`. Other vegetation paramaters are assumed time-invarient.
}
\examples{
# Run model:
micropoint<-runpointmodel(climdata,0.05,dtmcaerth,vegp,soilc)
# Plot canopy heat exchange surface temperature
microp<-micropoint$dfo
plot(microp$Tc,type="l") # temperature of canopy surface
}
