% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snowenv}
\alias{snowenv}
\title{A 2.5 degree resolution global dataset indicating snow environments}
\format{
A PackedSpatRaster object with 72 rows and 144 columns
}
\source{
derived from \url{https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world}
}
\usage{
snowenv
}
\description{
A spatial dataset of indicating snow environments (0 = Alpine, 1 = maritime, 2 = Prairie, 3 = Taiga, 4 = Tundra)
}
\keyword{datasets}
