% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilc}
\alias{soilc}
\title{A dataset of soil characteristics.}
\format{
A list of  the following objects:
\describe{
  \item{soiltype}{a RasterLayer object of numeric integers corresponding to soil types given in soilparameters}
  \item{groundr}{a RasterLayer object of soil reflectance of shortwave radiation}
}
}
\usage{
soilc
}
\description{
An object of class soilcharac for the area bounded by 169475, 169525, 12475, 12525
(xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference system (CRS: 27700).
}
\keyword{datasets}
