% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{soilcfromtype}
\alias{soilcfromtype}
\title{Create an object of type `soilcharac`}
\usage{
soilcfromtype(soiltype, groundr = 0.15)
}
\arguments{
\item{groundr}{a single numeric value, matrix or SpatRaster object of soil reflectance (in range 0-1)}

\item{a}{SpatRaster object of soiltypes numbered numerically as integers as in `soilparameters`}
}
\value{
an object of class `soilcharac`, a list of two SpatRaster object of `soiltype` and
`groundr` converted to a SpatRaster object.
}
\description{
The function `soilcfromtype` creates an object of class `soilcharac`
from a SpatRaster object of soil types and specified ground reflectance
}
