% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{soilmdistribute}
\alias{soilmdistribute}
\title{Spatially distribute soil moisture by topographic wetness index}
\usage{
soilmdistribute(soilm, dtm, Smin = 0.074, Smax = 0.422, tfact = 1.5, twi = NA)
}
\arguments{
\item{soilm}{a vector of soil moisture fractions}

\item{Smin}{minimum fractional soil moisture content}

\item{Smax}{maximum fractional soil moisture content}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness}

\item{twi}{optional raster of topographic wetness index values.
Calculated from `dtm` of not supplied, but outer cells will be NA.}

\item{a}{digital terrain model (used for calaculating topographic wetness)}
}
\value{
a 3D array of soil moistures with the same x and y dims as `dtm` and z
equivelent to length(soilm)
}
\description{
The function `soilmdistribute` spatially distrubutes soil moisture by the Bevan and Kirkby
topographic wetness index
}
\examples{
# Calculate vector of soil moistures
library(raster)
swrad <- (1 - 0.15) * climdata$swrad
lwout <- 5.67e-8 * 0.95 * (climdata$temp + 273.15)^4
lwnet <- (1 - climdata$skyem) * lwout
rnet <- swrad - lwnet
soilm <- soilmpredict(rainfall, rnet, "Loam", c(0.4, 0.4))$soilm1
sma<- soilmdistribute(soilm, dtmcaerth)
plot(raster(sma[,,365])) # soil moisture on last day of year
plot(raster(sma[,,188])) # soil moisture on driest day of year

}
