% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rimplementation.R
\name{soilmdistribute}
\alias{soilmdistribute}
\title{Spatially distribute soil moisture by topographic wetness index}
\usage{
soilmdistribute(micro, tfact = 1.5)
}
\arguments{
\item{micro}{an object of class microin as returned by e.g. modelin}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness}
}
\value{
a 3D array of soil moistures with the same x and y dims as `dtm` and z
equivelent to length(soilm)
}
\description{
The function `soilmdistribute` spatially distrubutes soil moisture by the Bevan and Kirkby
topographic wetness index
}
\examples{
micropoint <- runpointmodel(climdata, 0.05, dtmcaerth, vegp, soilc)
micropoint <- subsetpointmodel(micropoint)
micro <- modelin(micropoint, vegp, soilc, dtmcaerth)
micro <- soilmdistribute(micro) # default tfact
plot(rast(micro$soilm[,,1]))
}
