% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{soiltemp_dy}
\alias{soiltemp_dy}
\title{Calculates ground surface temperature (daily)}
\usage{
soiltemp_dy(
  micro_dy,
  reqhgt = 0.05,
  pai_a = NA,
  tfact = 1.5,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA,
  twi = NA
)
}
\arguments{
\item{micro_dy}{an object of class `microindaily` as returned by [modelin_dy()] (see details)}

\item{reqhgt}{height above ground at which model outputs are needed (m).}

\item{pai_a}{plant area index above `reqhgt`. Determined from total `pai` if not supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{slr}{an optional SpatRast object of slope values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional SpatRast object of aspect values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from dtm if not supplied, but outer cells will be NA.}

\item{twi}{optional SpatRast object of topographic wetness index values.
Calculated from `dtm` of not supplied, but outer cells will be NA.}
}
\value{
an object of class microdaily with the following components added:

`T0` ground surface temperature (deg C)

`G` ground heat flux (W/m^2)

additional terms needed for subsequent modelling.
}
\description{
The function `soiltemp_dy` estimates ground surface temperature
}
\details{
if [wind()] has not been run to add additional elements to `micro`
it is automatically called.
}
\seealso{
[wind()], [soiltemp_hr()]
}
