% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{temphumE}
\alias{temphumE}
\title{Estimate temperature and humidity at specified height above ground}
\usage{
temphumE(
  micro,
  reqhgt = 0.05,
  pai_a = NA,
  tfact = 1.5,
  surfwet = NA,
  slr = NA,
  apr = NA,
  hor = NA,
  twi = NA,
  wsa = NA
)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{reqhgt}{height above ground for which wind speeds are wanted. If negative (below ground) wind friction velocity only is returned}

\item{pai_a}{plant area index above `reqhgt`. Determined from total `pai` if not supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{surfwet}{an optional coefficient describing the fraction of the
vegetation surface that acts like a free water surface. Calculated from rainfall if not supplied.}

\item{slr}{an optional SpatRaster object of slope values (Radians). Calculated from
the dtm in micro if not supplied, but outer cells will be NA.}

\item{apr}{an optional SpatRaster object of aspect values (Radians). Calculated from
the dtm in micro if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from the dtm in micro if not supplied, but outer cells will be NA.}

\item{twi}{optional SpatRast object of topographic wetness index values.
Calculated from the dtm in micro if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from the dtm in micro if not supplied, but outer cells will be NA.}
}
\value{
`Tz` Array of air temperatures at height `reqhgt` (deg C)

`tleaf` Array of leaf temperatures at height `reqhgt` (deg C). if `reqhgt` greater than canopy height, the returned temperatures are those that would be attained by an unshaded leaf with average reflectance

`T0` Array of ground surface temperatures (deg C)

`relhum` Array of relative humidities at height `reqhgt` (percentage)

`windspeed` Array of wind speeds at height `reqhgt` (m/s)

`Rdirdown` Array of downward direct shortwave radiation received on horizontal surface at height `reqhgt` (W/m^2 - per unit one-sided area)

`Rdifdown` Array of downward diffuse shortwave radiation received on horizontal surface at height `reqhgt` (W/m^2 - per unit one-sided area)

`Rlwdown` Array of downward longwave radiation received on horizontal surface at height `reqhgt` (W/m^2 - per unit one-sided area)

`Rswup` Array of upward shortwave radiation received on the underside of a horizontal surface at height `reqhgt` (W/m^2 - per unit one-sided area). Upward shortwave radiation is assumed to be diffuse.

`Rlwup` Array of upward longwave radiation received on the underside of a horizontal surface at height `reqhgt` (W/m^2 - per unit one-sided area).
}
\description{
The function `temphumE` runs the above ground component of
the microclimate model.
}
\details{
`pai_a` is used to calculate the radiation intercepted by leaves at `reqhgt` if
below canopy. If not supplied it is calculated from total plant area index by
assuming leaf density within the canopy is uniformly vertically distributed. If supplied
it must have the same dimensions as micro$pai. I.e. with the same x and y dims as the
the supplied dtm and values for each hour as the z dimension. The parameter `surfwet`
determines how much of the canopy should be treated as wet surface when calaculating
latent heat fluxes. However, except when extremely droughted, the matric potential of leaves
is such that `surfwet` ~ 1.
}
\seealso{
[below_hr()] and [below_day()] for running the below ground components of the microclimate model
}
