% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{temphumE}
\alias{temphumE}
\title{Estimate temperature and humidity at specified height above ground}
\usage{
temphumE(
  micro,
  reqhgt,
  pai_a = NA,
  folden = NA,
  xyf = 1,
  zf = NA,
  soilinit = c(NA, NA),
  tfact = 1.5,
  surfwet = 1,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA,
  maxhgt = NA,
  twi = NA
)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{reqhgt}{height above ground at which model outputs are needed (m). Must be positive}

\item{pai_a}{an optional array of plant area index values above `reqhgt` (see details)}

\item{folden}{an optional array of foliage densities at height `reqhgt` (m^3/m^3)}

\item{xyf}{optional input for called function [wind()]}

\item{zf}{optional input for called function [wind()]}

\item{soilinit}{initial soil moisture fractions in surface and subsurface layer (see [soilmpredict()])}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{surfwet}{an optional single numeric value of array of values specifying the proportion
of the canopy surface that should be treated as wet surface (see details)}

\item{slr}{an optional RasterLayer of slope values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional RasterLayer of aspect values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from dtm if not supplied, but outer cells will be NA.}

\item{maxhgt}{an optional height (m) for which wind speed is needed. Determined
from height of tallest vegetation or as 2 m, whichever is greater, if not supplied.}

\item{twi}{optional raster of topographic wetness index values.
Calculated from `dtm` of not supplied, but outer cells will be NA.}
}
\value{
an object of class microout with the following components:

`Tz` Array of air temperatures at height `reqhgt` (deg C)

`tleaf` Array of leaf temperatures at height `reqhgt` (deg C). NA if `reqhgt` greater than canopy height

`T0` Array of ground surface temperatures (deg C)

`relhum` Array of relative humidities at height `reqhgt` (percentage)

`windspeed` Array of wind speeds at height `reqhgt` (m/s)

`raddir` Array of direct shortwave radiation received on horizontal surface (W/m^2)

`raddif` Array of direct shortwave radiation received on horizontal surface (W/m^2)

`radlw` Array of downward longwave radiation received on horizontal surface (W/m^2)
}
\description{
The function `temphumE` runs the above ground component of
the microclimate model.
}
\details{
`pai_a` is used to calaculate the radiation incercepted by leaves at `reqhgt` if
below canopy. If not supplied it is calaculated from total plant area index by
assuming leaf density within the canopy is uniformly vertically distributed. If suuplied
it must have the same dimensions as micro$pai. I.e. with the same x and y dims as the
the supplied dtm and values for each hour as the z dimension. The paramater `surfwet`
determines how much of the canopy should be treated as wet surface when calaculating
latent heat fluxes. However, except when extremely droughted, the matric potential of leaves
is such that `surfwet` ~ 1.
}
\seealso{
[below_hr()] and [below_day()] for running the below ground components of the microclimate model
}
