% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegp}
\alias{vegp}
\title{A dataset of vegetation parameters.}
\format{
A list of  the following objects:
\describe{
  \item{pai}{an array of monthly plant area index values}
  \item{hgt}{a PackedSpatRaster object of vegetation heights (m)}
  \item{x}{a PackedSpatRaster object of the ratio of vertical to horizontal projections of leaf foliage}
  \item{gsmax}{a PackedSpatRaster object maximum stomatal conductances (mol / m^2 / s)}
  \item{leafr}{a PackedSpatRaster object of leaf reflectance values}
  \item{clump}{a PackedSpatRaster object of values between 0 and 1 indicating the degree of canopy clumpiness}
  \item{leafd}{a PackedSpatRaster object of leaf diameters (m)}
}
}
\usage{
vegp
}
\description{
An object of class vegparams for the area for the area bounded by 169475, 169525, 12475, 12525
(xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference system (CRS: 27700).
}
\keyword{datasets}
