% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegfromhabitat.R
\name{vegpfromhab}
\alias{vegpfromhab}
\title{Generate vegpetation paramaters form habitat type}
\usage{
vegpfromhab(habitats, hgts = NA, pai = NA, lat, long, tme)
}
\arguments{
\item{habitats}{a data.frame of weather variables (see details)}

\item{hgts}{an optional raster of vegetation heights. Estimated from habitat type if not provided.}

\item{pai}{an optional array of plant area index values. Estimated at monthly intervals
from habitat type, with seasonal variation dtermined form location and date if not provided.}

\item{lat}{latitude in decimal degrees. Only needed if `pai` not provided.}

\item{long}{longitude in decimal degrees. Only needed if `pai` not provided.}

\item{tme}{POSIXlt object of dates. Only needed if `pai` not provided (see details).}
}
\value{
an object of class vegparams - a list with the following elements:

`pai` an array of monthly plant area index values (see details).

`hgt` a raster if vegetation heights (m)

`x` a raster of ratios of vertical to horizontal projections of leaf foliage

`gsmax` a raster of maximum stomatal conductances (mol / m^2 / s)

`leafr` a raster of leaf reflectance values (to shortwave radiation)

`clump` a raster indicating the degree of canopy clumpiness, here set to 0 (vegetation not clumped)

`leafd` a raster of mean leaf widths (m)
}
\description{
The function `vegpfromhab` generates an object of class vegparams from
a RasterLayer object of habitat types
}
\details{
This function estimates the vegetation parameters needed to run microclimf from habitat types.
Plant area index values represent the combined one sided woody and green vegetation
plant area per unit ground area. If not provided, then approximated
from habitat type, location and date. The procedure is based on calibration
against MODIS-derived estimates, accounting for regional climate. An inbuilt dataset
of regional rainfall and temperature is included with the package. Monthly
values for each unique month in `tme` are returned. Note that values are
assumed spatially constant across a given habitat type, which is unlikely
to be the case in reality. Likewise, if vegetation height values are not provided,
these are estimated form habitat type and assumed constant within that habitat type.
Habitat types should be expressed as integers as follows:
(1) for Evergreen needleleaf forest,
(2) for Evergreen broadleaf forest,
(3) for Deciduous needleleaf forest,
(4) for Deciduous broadleaf forest,
(5) for Mixed forest,
(6) for Closed shrubland,
(7) for Open shrubland,
(8) for Woody savanna,
(9) for Savanna,
(10) for Short grassland,
(11) for Tall grassland,
(12) for Permanent wetland,
(13) for Cropland,
(14) for Urban and built-up,
(15) for Cropland / Natural vegetation mosaic and
(16) for Barren or sparsely vegetated
}
\examples{
library(raster)
tme<-as.POSIXlt(c(0:8783)*3600,origin="2000-01-01 00:00", tz = "GMT")
veg<-vegpfromhab(habitats,lat=50,long=-5,tme=tme)
plot(raster(veg$pai[,,1]), main = "Jan PAI")
plot(veg$hgt, main = "Vegetation height")
plot(veg$x, main = "Leaf angle coefficient")
plot(vegp$gsmax, main = "Maximum stomatal conductance")
plot(veg$leafr, main = "Leaf reflectance")
plot(veg$clump, main = "Canopy clumping factor")
plot(veg$leafd, main = "Leaf diamater")
}
