% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{wind}
\alias{wind}
\title{Downscales wind}
\usage{
wind(
  micro,
  reqhgt,
  pai_a = NA,
  xyf = 1,
  zf = NA,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA,
  maxhgt = NA
)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{reqhgt}{height above ground for which wind speeds are wanted. If negative (below ground) wind friction velocity only is returned}

\item{pai_a}{plant area index above `reqhgt`. Determined from total `pai` if not supplied.}

\item{xyf}{optional spatial smoothing factor applied in calculation of wind speeds (see details)}

\item{zf}{optional integer used to calculate how frequently to sample the plant area index array
when calculating smoothed wind speed (see details)}

\item{slr}{an optional SpatRaster object of slope values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional SpatRaster object of aspect values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from dtm if not supplied, but outer cells will be NA.}

\item{maxhgt}{an optional height (m) for which wind speed is needed. Determined
from height of tallest vegetation or as 2 m, whichever is greater, if not supplied.}
}
\value{
an object of class microin with the following components added:

`uf` wind fricion velocities (m/s)

`uz` optionally wind speed at height `reqhgt` (m/s)

aditional terms required for subsequent modelling
}
\description{
The function `wind` downscales wind speed accounting for
vegetation and terrain
}
\details{
In downscaling wind, two processes are accounted for. Firstly the drag effects
of vegetation on wind , which ultimately dictate the wind height profile. In
so doing, it is necessary to accommodate the fact the wind speed is not just affected by
the surface roughness at each point location, but also by vegetation surrounding the
location. This is accommodated for by applying `xyf` which effectively smooths though surface roughness
using `terra::aggregate` where `xyf` is the aggregation factor. Because surface roughness
also depends on the plant area index (PAI), which is time-variant, and is computationally intensive
to perform aggregations on PAI at each hourly time increment, `zf` determines the time at
what frequency to aggregate PAI. E.g. where `zf` is 100, PAI at every 100th time increment is aggregated.
By default 'xyf' is set to 1 and `zf` to aggregate PAI at monthly intervals. A second complication arises in that the maximum
vegetation height within the study area may exceed the height at which input wind speeds are
measured (by default assumed measured at a height of 2 m). The issue here is that wind speed
measurements are normally taken in open areas where wind is unimpeded by vegetation.
For this reason, the wind speed at `maxhgt` is calculated
using a standard height profile adjustment for an open area typical of those at which weather stations are sighted
prior to adjusting for the effects of local vegetation and terrain on wind speed. Terrain effects
are calculated by applying the topographic shelter coefficient described in Maclean et al
(2019) Methods in Ecology and Evolution, 10:280-290.
}
