% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{wind}
\alias{wind}
\title{Downscales wind}
\usage{
wind(
  micro,
  reqhgt = 0.05,
  pai_a = NA,
  tfact = 1.5,
  slr = NA,
  apr = NA,
  hor = NA,
  twi = NA,
  wsa = NA
)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{reqhgt}{height above ground for which wind speeds are wanted. If negative (below ground) wind friction velocity only is returned}

\item{pai_a}{plant area index above `reqhgt`. Determined from total `pai` if not supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see [soilmdistribute()])}

\item{slr}{an optional SpatRaster object of slope values (Radians). Calculated from the
dtm in micro if not supplied, but outer cells will be NA.}

\item{apr}{an optional SpatRaster object of aspect values (Radians). Calculated from
the dtm ion micro if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from the dtm in micro if not supplied, but outer cells will be NA.}

\item{twi}{optional SpatRast object of topographic wetness index values.
Calculated from the dtm in micro if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from the dtm in micro if not supplied, but outer cells will be NA.}
}
\value{
an object of class microin with the following components added:

`uf` wind friction velocity (m/s)

`uz` if `reqhgt > 0` wind speed at height `reqhgt` (m/s)

`gHa` convective conductance (mol/m^2/s)
}
\description{
The function `wind` downscales wind speed accounting for
vegetation and terrain
}
\details{
In downscaling wind, two processes are accounted for. Firstly the drag effects
of vegetation on wind , which ultimately dictate the wind height profile. Secondly,
the effects of local vegetation and terrain on wind speed. Terrain effects are calculated
by applying the topographic shelter coefficient described in Maclean et al (2019) Methods
in Ecology and Evolution, 10:280-290.
}
