% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentfunctions.R
\name{wind}
\alias{wind}
\title{Downscales wind}
\usage{
wind(
  micro,
  xyf = 1,
  zf = NA,
  psi_m = 0,
  reqhgt = NA,
  slr = NA,
  apr = NA,
  hor = NA,
  wsa = NA,
  maxhgt = NA
)
}
\arguments{
\item{micro}{object of class microin as returned by [modelin()]}

\item{xyf}{optional spatial smoothing factor applied in calculation of wind speeds (see details)}

\item{zf}{optional integer used to calculate how frequently to sample the plant area index array
when calculating smoothed wind speed (see details)}

\item{psi_m}{an optional diabatic correction factor accounting for the strength of surface heating}

\item{reqhgt}{option height above ground for which wind speeds are wanted. If ommited only the surface wind
frictian velocity is returned.}

\item{slr}{an optional SpatRaster object of slope values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{apr}{an optional SpatRaster object of aspect values (Radians). Calculated from
dtm if not supplied, but outer cells will be NA.}

\item{hor}{an optional array of the tangent of the angle to the horizon in
24 directions. Calculated from dtm if not supplied, but outer cells will be NA.}

\item{wsa}{an optional array of wind shelter coefficients in 8 directions.
Calculated from dtm if not supplied, but outer cells will be NA.}

\item{maxhgt}{an optional height (m) for which wind speed is needed. Determined
from height of tallest vegetation or as 2 m, whichever is greater, if not supplied.}
}
\value{
an object of class microin with the following components added:

`uf` wind fricion velocities (m/s)

`uz` optionally wind speed at height `reqhgt` (m/s)

aditional terms required for subsequent modelling
}
\description{
The function `wind` downscales wind speed accounting for
vegetation and terrain
}
\details{
In downscaling wind, two processes are accounted for. Firstly the drag effects
of vegetation on wind , which ultimately dictate the wind height profile. In
so doing, it is necessary to accomodate the fact the wind speed is not just affected by
the surface roughness at each point location, but also by vegetation surrounding the
location. This is accomodated for by applying `xyf` which effectively smooths though surface roughness
using `terra::aggregate` where `xyf` is the aggregation factor. Because surface roughness
also depends on the plant area index (PAI), which is time-varient, and is computationally intensive
to perform aggregations on PAI at each hourly time increment, `zf` determines the time at
what frequency to aggregate PAI. E.g. where `zf` is 100, PAI at every 100th time increment is aggregated.
By default 'xyf' is set to 10 m or the pixel resolution, which ever is greater and `zf`
is set to aggregate PAI at monthly intervals. A second complication arises in that the maximum
vegetation height witin the study area may exceed the height at which input wind speeds are
measured (by default assumed measured at a height of 2 m). For this reason, the wind
speed at the maximum height of vegetation within the study is calaculated using a standard height
profile adjustment for an open area typical of those at which weather stations are sighted
prior to adjusting for the effects of local vegetation and terrain on wind speed. Terrain effects
are calaculated by applying the topographic shelter coefficient described in Maclean et al
(2019) Methods in Ecology and Evolution, 10:280-290.
}
